% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{arrange.tbl_ts}
\alias{arrange.tbl_ts}
\alias{arrange.grouped_ts}
\title{Arrange rows by variables}
\usage{
\method{arrange}{tbl_ts}(.data, ...)

\method{arrange}{grouped_ts}(.data, ..., .by_group = FALSE)
}
\arguments{
\item{.data}{A \code{tbl_ts}.}

\item{...}{A set of unquoted variables, separated by commas.}

\item{.by_group}{\code{TRUE} will sort first by grouping variables.}
}
\description{
Arrange rows by variables
}
\details{
If not arranging key and index in ascending order, a warning is
likely to be issued.
}
\seealso{
\link[dplyr:arrange]{dplyr::arrange}

\link[dplyr:arrange]{dplyr::arrange}
}
