% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tsibble.R
\name{as_tibble.tbl_ts}
\alias{as_tibble.tbl_ts}
\title{Coerce to a tibble or data frame}
\usage{
\method{as_tibble}{tbl_ts}(x, ...)
}
\arguments{
\item{x}{A \code{tbl_ts}.}

\item{...}{Ignored.}
}
\description{
Coerce to a tibble or data frame
}
\examples{
as_tibble(pedestrian)
}
