% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{as.tsibble}
\alias{as.tsibble}
\title{Deprecated functions}
\usage{
as.tsibble(x, ...)
}
\arguments{
\item{x}{Other objects.}
}
\description{
\itemize{
\item \code{is.tsibble()} \lifecycle{soft-deprecated}
\item \code{as.tsibble()} \lifecycle{deprecated}
\item \code{id()} \lifecycle{deprecated}
\item \code{fill_na()} \lifecycle{defunct}
}
}
\keyword{internal}
