% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{as_period}
\alias{as_period}
\title{Number of units between a sequence of index objects}
\usage{
as_period(x)
}
\arguments{
\item{x}{An index object including "yearmonth", "yearquarter", "Date" and others.}
}
\description{
Number of units between a sequence of index objects
}
\examples{
# at two months interval ----
x <- yearmth(seq(2016, 2018, by = 0.5))
as_period(x)
}
