% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-na.R
\name{case_na}
\alias{case_na}
\title{A thin wrapper of \code{dplyr::case_when()} if there are \code{NA}s}
\usage{
case_na(formula)
}
\arguments{
\item{formula}{A two-sided formula. The LHS expects a vector containing \code{NA},
and the RHS gives the replacement value.}
}
\description{
A thin wrapper of \code{dplyr::case_when()} if there are \code{NA}s
}
\examples{
x <- rnorm(10)
x[c(3, 7)] <- NA_real_
case_na(x ~ 10)
case_na(x ~ mean(x, na.rm = TRUE))
}
