% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{default_time_units}
\alias{default_time_units}
\title{Time units from tsibble's "interval" class used for \code{seq(by = )}}
\usage{
default_time_units(x)
}
\arguments{
\item{x}{An interval.}
}
\description{
\lifecycle{experimental}
}
\keyword{internal}
