% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{as.tsibble}
\alias{as.tsibble}
\alias{fill_na}
\alias{id}
\title{Deprecated functions}
\usage{
as.tsibble(x, ...)

fill_na(.data, ..., .full = FALSE)

id(...)
}
\arguments{
\item{x}{Other objects.}

\item{...}{Variables passed to tsibble()/as_tsibble().}
}
\description{
\itemize{
\item \code{as.tsibble()} \lifecycle{deprecated}
\item \code{id()} \lifecycle{deprecated}
\item \code{fill_na()} \lifecycle{defunct}
}
}
\keyword{internal}
