% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-wise.R
\name{diff}
\alias{diff}
\title{Lagged Differences in C++}
\usage{
diff(x, lag = 1, differences = 1)
}
\arguments{
\item{x}{A numeric vector.}

\item{lag}{An integer indicating which lag to use.}

\item{differences}{An integer indicating the order of the difference.}
}
\value{
A numeric vector of the same size of \code{x} with \code{NA} padding in.
}
\description{
Lagged Differences in C++
}
\examples{
# examples from base
diff(1:10, 2)
diff(1:10, 2, 2)
x <- cumsum(cumsum(1:10))
diff(x, lag = 2)
diff(x, differences = 2)
}
