% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-na.R
\name{fill_na}
\alias{fill_na}
\alias{fill_na.tbl_ts}
\title{Turn implicit missing values into explicit missing values}
\usage{
fill_na(.data, ...)

\method{fill_na}{tbl_ts}(.data, ...)
}
\arguments{
\item{.data}{A \code{tbl_ts}.}

\item{...}{A set of name-value pairs. The values will replace existing explicit
missing values by variable, otherwise \code{NA}.}
}
\description{
Turn implicit missing values into explicit missing values
}
\examples{
# not replace NA
fill_na(pedestrian)

# replace NA
pedestrian \%>\% 
  fill_na(
    Date = lubridate::as_date(Date_Time),
    Time = lubridate::hour(Date_Time),
    Count = median(Count, na.rm = TRUE)
  )
}
