% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile.R
\name{future_tile}
\alias{future_tile}
\title{Tiling window in parrallel}
\description{
Multiprocessing equivalents of \code{slide()}, \code{tile()}, \code{stretch()} with \code{future_} prefixed to them.
\itemize{
\item \code{future_*_lgl()}, \code{future_*_int()}, \code{future_*_dbl()}, \code{future_*_chr()},
\code{future_*_dfr()}, \code{future_*_dfc()}.
\item Extra arguments \code{.progress} and \code{.options} for enabling progress bar and the
future specific options to use with the workers.
}
}
\alias{future_tile}
\alias{future_tile2}
\alias{future_ptile}
\alias{future_tile_lgl}
\alias{future_tile_chr}
\alias{future_tile_int}
\alias{future_tile_dbl}
\alias{future_tile_dfr}
\alias{future_tile_dfc}
\alias{future_tile2_lgl}
\alias{future_tile2_chr}
\alias{future_tile2_int}
\alias{future_tile2_dbl}
\alias{future_tile2_dfr}
\alias{future_tile2_dfc}
\alias{future_ptile_lgl}
\alias{future_ptile_chr}
\alias{future_ptile_int}
\alias{future_ptile_dbl}
\alias{future_ptile_dfr}
\alias{future_ptile_dfc}
