% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{group_by_key}
\alias{group_by_key}
\title{Group by key variables}
\usage{
group_by_key(.data, ..., .drop = key_drop_default(.data))
}
\arguments{
\item{.data}{A \code{tbl_ts} object.}

\item{...}{Ignored.}

\item{.drop}{Drop groups formed by factor levels that don't appear in the
data? The default is \code{TRUE} except when \code{.data} has been previously
grouped with \code{.drop = FALSE}. See \code{\link[dplyr:group_by_drop_default]{group_by_drop_default()}} for details.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
tourism \%>\%
  group_by_key()
}
