% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{guess_frequency}
\alias{guess_frequency}
\title{Guess a time frequency from other index objects}
\usage{
guess_frequency(x)
}
\arguments{
\item{x}{An index object including "yearmonth", "yearquarter", "Date" and others.}
}
\description{
A possible frequency passed to the \code{ts()} function
}
\details{
If a series of observations are collected more frequently than
weekly, it is more likely to have multiple seasonalities. This function
returns a frequency value at its nearest ceiling time resolution. For example,
hourly data would have daily, weekly and annual frequencies of 24, 168 and 8766
respectively, and hence it gives 24.
}
\examples{
x <- yearmth(seq(2016, 2018, by = 0.25))
guess_frequency(x)
}
\references{
\url{https://robjhyndman.com/hyndsight/seasonal-periods/}
}
