% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holiday.R
\name{holiday_aus}
\alias{holiday_aus}
\title{Australia public holidays}
\usage{
holiday_aus(year, state = "National")
}
\arguments{
\item{year}{A vector of integer(s) indicating year(s).}

\item{state}{A state in Australia including "ACT", "NSW", "NT", "QLD", "SA",
"TAS", "VIC", "WA", as well as "National".}
}
\value{
A tibble consisting of \code{holiday} labels and their associated dates
in the year(s).
}
\description{
Australia public holidays
}
\details{
Not documented public holidays:
\itemize{
\item AFL public holidays for Victoria
\item Adelaide Cup for Southern Australia
\item Queen's Birthday for Western Australia
}
}
\examples{
holiday_aus(2016, state = "VIC")
holiday_aus(2013:2016, state = "ACT")
}
\references{
\href{https://www.australia.gov.au/about-australia/special-dates-and-events/public-holidays}{Public holidays}
}
