% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tsibble.R
\name{id}
\alias{id}
\title{Identifiers used for creating key}
\usage{
id(...)
}
\arguments{
\item{...}{Variables passed to tsibble()/as_tsibble().}
}
\description{
Identifiers used for creating key
}
\seealso{
\link{tsibble}, \link{as_tsibble}
}
