% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tsibble.R
\name{interval}
\alias{interval}
\alias{index}
\title{Return index and interval from a tsibble}
\usage{
interval(x)

index(x)
}
\arguments{
\item{x}{A tsibble object.}
}
\description{
Return index and interval from a tsibble
}
\examples{
data(pedestrian)
index(pedestrian)
interval(pedestrian)
}
