% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-sum.R
\name{index_sum}
\alias{index_sum}
\title{Extensible index type to tsibble}
\usage{
index_sum(x)
}
\arguments{
\item{x}{An object to be added to index types that tsibble supports.}
}
\description{
Extensible index type to tsibble
}
