% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tsibble.R
\name{is_tsibble}
\alias{is_tsibble}
\alias{is.tsibble}
\alias{is_grouped_ts}
\alias{is.grouped_ts}
\title{Test if the object is a tsibble}
\usage{
is_tsibble(x)

is.tsibble(x)
}
\arguments{
\item{x}{An object.}
}
\value{
TRUE if the object inherits from the tbl_ts class.
}
\description{
Test if the object is a tsibble
}
