% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsibble-meta.R
\name{key_data}
\alias{key_data}
\alias{key_rows}
\alias{key_size}
\alias{n_keys}
\title{Key metadata}
\usage{
key_data(.data)

key_rows(.data)

key_size(x)

n_keys(x)
}
\arguments{
\item{.data}{A tsibble}
}
\description{
Key metadata
}
\examples{
key_data(pedestrian)
}
\seealso{
\link[dplyr:group_data]{dplyr::group_data}
}
