% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tsibble.R
\name{key}
\alias{key}
\alias{key_vars}
\title{Return key variables}
\usage{
key(x)

key_vars(x)
}
\arguments{
\item{x}{A data frame.}
}
\description{
\code{key()} returns a list of symbols; \code{key_vars()} gives a character vector.
}
\examples{
# A single key for pedestrian data ----
data(pedestrian)
key(pedestrian)
key_vars(pedestrian)

# Nested and crossed keys for tourism data ----
data(tourism)
key(tourism)
key_vars(tourism)
}
