% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key-group.R
\name{key_by}
\alias{key_by}
\title{Change key variables for a given \code{tbl_ts}}
\usage{
key_by(.data, ...)
}
\arguments{
\item{.data}{A \code{tbl_ts}.}

\item{...}{Variables to construct the key.}
}
\description{
Change key variables for a given \code{tbl_ts}
}
\examples{
# By removing `State` from key, it is a valid tsibble too.
tourism \%>\%
  key_by(Region, Purpose)
}
