% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-sum.R
\name{key_sum}
\alias{key_sum}
\title{Summary of key variables}
\usage{
key_sum(x)
}
\arguments{
\item{x}{An object that contains "key".}
}
\description{
Summary of key variables
}
\examples{
key_sum(pedestrian)
}
