% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-wise.R
\name{keyed_lag}
\alias{keyed_lag}
\alias{keyed_lead}
\alias{keyed_difference}
\title{Tsibble-aware lag, lead, and difference operations}
\usage{
keyed_lag(var, n = 1L, default = NA)

keyed_lead(var, n = 1L, default = NA)

keyed_difference(var, lag = 1, differences = 1, default = NA)
}
\arguments{
\item{var}{Unquoted variable.}

\item{n}{a positive integer of length 1, giving the number of positions to
lead or lag by}

\item{default}{value used for non-existent rows. Defaults to \code{NA}.}

\item{lag}{An positive integer indicating which lag to use.}

\item{differences}{An positive integer indicating the order of the difference.}
}
\description{
They are specialist vector functions for tsibble, used in conjunction with
\code{mutate()}/\code{transform()}. Unlike their counterparts, such as \code{lag()} and \code{lead()},
\verb{keyed_*()} \emph{take care of} temporal ordering and gaps, and do the right thing.
}
\examples{
library(dplyr)
yrmth <- c("2018-07", "2018-08", "2018-11", "2018-12", "2019-01", "2019-03")
tsbl <- tsibble(
  yrmth = yearmonth(yrmth), 
  income = c(1153, 1181, 1236, 1297, 1264, 1282),
  index = yrmth)

tsbl \%>\% 
  mutate(
    lag_income = keyed_lag(income),
    lead_income = keyed_lead(income),
    diff_income = keyed_difference(income)
  )

# aware of key and gaps
pedestrian \%>\% 
  mutate(l_count = keyed_lag(Count))

# to replace the following chain
pedestrian \%>\% 
  fill_gaps() \%>\% 
  group_by_key() \%>\% 
  mutate(l_count = lag(Count))
}
