% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{lslide}
\alias{lslide}
\alias{lslide_if}
\alias{lslide_at}
\title{Rolling window on a list}
\usage{
lslide(.x, .f, ..., .size = 1, .fill = NA)

lslide_if(.x, .p, .f, ..., .size = 1, .fill = NA)

lslide_at(.x, .at, .f, ..., .size = 1, .fill = NA)
}
\arguments{
\item{.x}{A list or data frame.}

\item{.f}{A function that takes and returns a list or data frame.}

\item{...}{Additional arguments passed on to \code{.f}.}

\item{.size}{An integer for window size.}

\item{.fill}{A single value or data frame to replace \code{NA}.}

\item{.p}{A single predicate function, a formula describing such a
predicate function, or a logical vector of the same length as \code{.x}.
Alternatively, if the elements of \code{.x} are themselves lists of
objects, a string indicating the name of a logical element in the
inner lists. Only those elements where \code{.p} evaluates to
\code{TRUE} will be modified.}

\item{.at}{A character vector of names or a numeric vector of
positions. Only those elements corresponding to \code{.at} will be
modified.}
}
\description{
Rolling window on a list
}
