% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tsibble.R
\name{measured_vars}
\alias{measured_vars}
\title{Return measured variables}
\usage{
measured_vars(x)
}
\arguments{
\item{x}{A \code{tbl_ts}.}
}
\description{
Return measured variables
}
\examples{
measured_vars(pedestrian)
measured_vars(tourism)
}
