% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-data.R
\name{new_data}
\alias{new_data}
\alias{new_data.tbl_ts}
\alias{append_row}
\alias{append_case}
\title{New tsibble data and append new observations to a tsibble}
\usage{
new_data(.data, n = 1L, ...)

\method{new_data}{tbl_ts}(.data, n = 1L, keep_all = FALSE, ...)

append_row(.data, n = 1L, ...)
}
\arguments{
\item{.data}{A \code{tbl_ts}.}

\item{n}{An integer or \verb{[mixtime::duration]} indicates the sequence size
of new time indices to be generated. The sign of \code{n} determines the start
and direction of the sequence, with positive values generating indices
after the last time index and negative values generating indices before the
first time index of each key group.}

\item{...}{Passed to individual S3 method.}

\item{keep_all}{If \code{TRUE} keep all the measured variables as well as index
and key, otherwise only index and key.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{append_row()}: add new rows to the start/end of a tsibble by filling a key-index
pair and \code{NA} for measured variables.

\code{append_case()} is an alias of \code{append_row()}.
}
\examples{
new_data(pedestrian)
new_data(pedestrian, keep_all = TRUE)
new_data(pedestrian, n = 3)
new_data(pedestrian, n = -2)

new_data(pedestrian, n = mixtime::days(1L))

tsbl <- tsibble(
  date = rep(as.Date("2017-01-01") + 0:2, each = 2),
  group = rep(letters[1:2], 3),
  value = rnorm(6),
  key = group
)
append_row(tsbl)
append_row(tsbl, n = 2)
append_row(tsbl, n = -2)
}
