% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-data.R
\name{new_data}
\alias{new_data}
\title{New tsibble data}
\usage{
new_data(.data, n = 1L, .keep_all = FALSE)
}
\arguments{
\item{.data}{A \code{tbl_ts}.}

\item{n}{An integer indicates the number of key-index pair to append.}

\item{.keep_all}{Keep all the measured variables.}
}
\description{
New tsibble data
}
\examples{
new_data(pedestrian)
new_data(pedestrian, .keep_all = TRUE)
new_data(pedestrian, n = 3)
}
\seealso{
\link{append_row} for appending new observations to a tsibble
}
