% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{new_interval}
\alias{new_interval}
\title{Interval constructor for a tsibble}
\usage{
new_interval(..., .regular = TRUE, .others = list())
}
\arguments{
\item{...}{A set of name-value pairs to specify default interval units: "year",
"quarter", "month", "week", "day", "hour", "minute", "second", "millisecond",
"microsecond", "nanosecond", "unit".}

\item{.regular}{Logical. \code{FALSE} gives an irregular interval, and will ignore
the \code{...} argument.}

\item{.others}{A list name-value pairs that are not included in the \code{...},
to allow custom interval.}
}
\value{
an "interval" class
}
\description{
\lifecycle{stable}
\code{new_interval()} creates an interval object.
}
\examples{
new_interval(hour = 1, minute = 30)
new_interval(.regular = FALSE) # irregular interval
new_interval() # unknown interval
new_interval(.others = list(semester = 1)) # custom interval
}
