% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pedestrian}
\alias{pedestrian}
\title{Pedestrian counts in Melbourne city}
\format{A tsibble with 66071 rows and 5 variables:
\describe{
\item{Sensor}{Sensor names (key variable)}
\item{Date_Time}{Date time when the pedestrian counts are recorded (index)}
\item{Date}{Date when the pedestrian counts are recorded}
\item{Time}{Hour associated with Date_Time}
\item{Counts}{Hourly pedestrian counts}
}}
\usage{
pedestrian
}
\description{
A dataset containing the hourly pedestrian counts from 2015-01-01 to
2016-12-31 at 4 sensors in the city of Melbourne. The variables are as follows:
}
\keyword{datasets}
