% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{yearquarter}
\alias{yearquarter}
\title{Represent year-week (ISO) starting on Monday, year-month or year-quarter objects}
\usage{
yearquarter(x)
}
\arguments{
\item{x}{Other object.}
}
\value{
Year-week (\code{yearweek}), year-month (\code{yearmonth}) or year-quarter
(\code{yearquarter}) objects.
}
\description{
\lifecycle{stable}

Create or coerce using \code{yearweek()}, \code{yearmonth()}, or \code{yearquarter()}
}
\section{Index functions}{

The tsibble \code{yearmonth()} and \code{yearquarter()} function respects time zones of
the input \code{x}, contrasting to their zoo counterparts.
}

\examples{
# coerce POSIXct/Dates to yearweek, yearmonth, yearquarter
x <- seq(as.Date("2016-01-01"), as.Date("2016-12-31"), by = "1 month")
yearweek(x)
yearmonth(x)
yearmonth(yearweek(x))
yearquarter(x)

# coerce yearmonths to yearquarter
y <- yearmonth(x)
yearquarter(y)

# parse characters
yearweek(c("2018 W01", "2018 Wk01", "2018 Week 1"))
yearmonth(c("2018 Jan", "2018-01", "2018 January"))
yearquarter(c("2018 Q1", "2018 Qtr1", "2018 Quarter 1"))

# seq() and binary operaters
wk1 <- yearweek("2017-11-01")
wk2 <- yearweek("2018-04-29")
seq(from = wk1, to = wk2, by = 2) # by two weeks
wk1 + 0:9
mth <- yearmonth("2017-11")
seq(mth, length.out = 5, by = 1) # by 1 month
mth + 0:9
seq(yearquarter(mth), length.out = 5, by = 1) # by 1 quarter

# different formats
format(c(wk1, wk2), format = "\%V/\%Y")
format(y, format = "\%y \%m")
format(yearquarter(mth), format = "\%y Qtr\%q")
}
\seealso{
\link{interval_pull}
}
