% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tsibble.R
\name{interval}
\alias{interval}
\alias{is_regular}
\alias{is_ordered}
\title{Meta-information of a tsibble}
\usage{
interval(x)

is_regular(x)

is_ordered(x)
}
\arguments{
\item{x}{A tsibble object.}
}
\description{
\itemize{
\item \code{interval()} returns an interval of a tsibble.
\item \code{is_regular} checks if a tsibble is spaced at regular time or not.
\item \code{is_ordered} checks if a tsibble is ordered by key and index.
}
}
\examples{
interval(pedestrian)
is_regular(pedestrian)
is_ordered(pedestrian)
}
