% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{arrange.tbl_ts}
\alias{arrange.tbl_ts}
\alias{filter.tbl_ts}
\alias{slice.tbl_ts}
\title{Row-wise verbs}
\usage{
\method{arrange}{tbl_ts}(.data, ...)

\method{filter}{tbl_ts}(.data, ...)

\method{slice}{tbl_ts}(.data, ...)
}
\arguments{
\item{.data}{A tsibble.}

\item{...}{\itemize{
\item A set of unquoted variables for \code{arrange()}.
\item Logical predicates defined in terms of the variables for \code{filter()}.
\item Integer row numbers for \code{slice()}.
}}
}
\description{
\code{arrange()} arranges rows by variable; \code{filter()} returns rows with matching
conditions; \code{slice()} selects rows by position.
}
\seealso{
\link[dplyr:arrange]{dplyr::arrange}

\link[dplyr:filter]{dplyr::filter}

\link[dplyr:slice]{dplyr::slice}
}
