% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaps.R
\name{scan_gaps}
\alias{scan_gaps}
\alias{scan_gaps.tbl_ts}
\title{Scan a tsibble for implicit missing observations}
\usage{
scan_gaps(.data, ...)

\method{scan_gaps}{tbl_ts}(.data, .full = FALSE, ...)
}
\arguments{
\item{.data}{A \code{tbl_ts}.}

\item{...}{Other arguments passed on to individual methods.}

\item{.full}{\code{FALSE} to find gaps for each group within its own period. \code{TRUE}
to find gaps over the entire time span of the data.}
}
\description{
Scan a tsibble for implicit missing observations
}
\examples{
scan_gaps(pedestrian)
}
\seealso{
Other implicit gaps handling: \code{\link{count_gaps}},
  \code{\link{fill_gaps}}, \code{\link{has_gaps}}
}
\concept{implicit gaps handling}
