% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{select.tbl_ts}
\alias{select.tbl_ts}
\alias{rename.tbl_ts}
\title{Select/rename variables by name}
\usage{
\method{select}{tbl_ts}(.data, ..., drop = FALSE)

\method{rename}{tbl_ts}(.data, ...)
}
\arguments{
\item{.data}{A tsibble.}

\item{...}{Unquoted variable names separated by commas. \code{rename()} requires
named arguments.}

\item{drop}{\code{FALSE} returns a tsibble object as the input. \code{TRUE} drops a
tsibble and returns a tibble.}
}
\description{
Select/rename variables by name
}
\details{
These column-wise verbs from dplyr have an additional argument of \code{drop = FALSE}
for tsibble. The index variable cannot be dropped for a tsibble. If any key
variable is changed, it will validate whether it's a tsibble internally.
Turning \code{drop = TRUE} converts to a tibble first and then do the operations.
}
\seealso{
\link[dplyr:select]{dplyr::select}

\link[dplyr:rename]{dplyr::rename}
}
