% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sincos}
\alias{sincos}
\title{Sythetic data as an example of hms/difftime}
\format{A tsibble with 360 rows and 3 variables:
\itemize{
\item \strong{length}: Time length as \code{hms} and \code{difftime} class (index)
\item \strong{trig}: Trigonometry as \emph{sine} and \emph{cosine} (key)
\item \strong{value}: Value
}}
\usage{
sincos
}
\description{
A dataset containing the values recorded per second from \code{00:00:00} to \code{00:02:59}.
}
\keyword{datasets}
