% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{slice.tbl_ts}
\alias{slice.tbl_ts}
\title{Selects rows by position}
\usage{
\method{slice}{tbl_ts}(.data, ...)
}
\arguments{
\item{.data}{A \code{tbl_ts}.}

\item{...}{Unique integers of row numbers to be selected.}
}
\description{
Selects rows by position
}
\details{
If row numbers are not in ascending order, a warning is likely to
be issued.
}
\seealso{
\link[dplyr:slice]{dplyr::slice}
}
