% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-key.R
\name{split_by}
\alias{split_by}
\alias{split_by.tbl_ts}
\title{Split data into a list by variables}
\usage{
split_by(x, ...)

\method{split_by}{tbl_ts}(x, ...)
}
\arguments{
\item{x}{A \code{tbl_ts}.}

\item{...}{Unquoted variables to split by.}
}
\description{
Split data into a list by variables
}
\examples{
pedestrian \%>\% 
  split_by(Sensor)
}
