% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-key.R
\name{split_key}
\alias{split_key}
\alias{split_key.tbl_ts}
\title{Split data into a list by its key}
\usage{
split_key(x)

\method{split_key}{tbl_ts}(x)
}
\arguments{
\item{x}{A \code{tbl_ts}.}
}
\description{
Split data into a list by its key
}
\examples{
split_key(pedestrian)
}
