% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{stretch}
\alias{stretch}
\alias{stretch.numeric}
\alias{stretch.data.frame}
\alias{stretcher}
\title{Stretching window calculation}
\usage{
stretch(x, .f, ..., size = 1, init = 1)

\method{stretch}{numeric}(x, .f, ..., size = 1, init = 1)

\method{stretch}{data.frame}(x, .f, ..., size = 1, init = 1,
  deframe = TRUE)

stretcher(x, size = 1, init = 1)
}
\arguments{
\item{x}{A vector of numerics, or data frame.}

\item{.f}{A function, formula, or atomic vector.

If a \strong{function}, it is used as is.

If a \strong{formula}, e.g. \code{~ .x + 2}, it is converted to a function. There
are three ways to refer to the arguments:
\itemize{
\item For a single argument function, use \code{.}
\item For a two argument function, use \code{.x} and \code{.y}
\item For more arguments, use \code{..1}, \code{..2}, \code{..3} etc
}

This syntax allows you to create very compact anonymous functions.

If \strong{character vector}, \strong{numeric vector}, or \strong{list}, it
is converted to an extractor function. Character vectors index by name
and numeric vectors index by position; use a list to index by position
and name at different levels. Within a list, wrap strings in \code{\link[=get-attr]{get-attr()}}
to extract named attributes. If a component is not present, the value of
\code{.default} will be returned.}

\item{...}{Additional arguments passed on to \code{.f}.}

\item{size, init}{An integer for moving and initial window size.}

\item{deframe}{TRUE a list is returned. FALSE returns a data frame.}
}
\description{
\code{stretch()} is an S3 method to carry out expanding window calculation;
\code{stretcher()} splits the input \code{x} to a list according to the window size.
}
\examples{
stretch(1:10, mean, init = 3)
stretcher(1:10, init = 3)
}
\seealso{
\link{slide} for sliding window with overlapping observations;
\link{tile} for tiling window without overlapping observations.
}
