% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsummerise.R
\name{tsummarise.tbl_ts}
\alias{tsummarise.tbl_ts}
\title{Aggregate over calendar periods}
\usage{
tsummarise.tbl_ts(.data, ...)
}
\arguments{
\item{.data}{A tsibble (of \code{tbl_ts} class).}

\item{...}{Name-value pairs of summary functions. To aggregate tsibble over
a certain calendar period, for example yearly aggregates, \code{~ year()} needs
passing to \code{...}. Please see details.}
}
\value{
A tsibble class when the \code{~} is present.
}
\description{
It computes summary statistics for a tsibble over calendar
periods, usually used in combination of \link{group_by}.
}
\details{
It's S3 method implemented for tsibble() (\code{tbl_ts}) obtained from
\code{\link[dplyr:summarise]{dplyr::summarise()}}. A formula with \code{~} followed by one of calendar component
functions from base, \link{lubridate} and \link{zoo} specifies the period when summary
functions are carried out.  Currently \code{~ year()} indicates yearly aggregates.
\code{~ yearqtr()} indicates quarterly aggregates. \code{~ yearmon()} indicates
monthly aggregates. \code{~ as_date()} or \code{as.Date()} indicates daily aggregates.
}
\examples{
   # pkgs_ts <- as_tsibble(tidypkgs, index = date, package)
   # pkgs_ts \%>\% 
   #   group_by(package) \%>\% 
   # summarise(avg_count = mean(count), month = ~ as.yearmon())

}
\seealso{
\link[dplyr:summarise]{dplyr::summarise}
}
\author{
Earo Wang
}
