% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaps.R
\name{tbl_gaps}
\alias{tbl_gaps}
\title{Find missing elements in \code{x} with respect to \code{y}}
\usage{
tbl_gaps(x, y)
}
\arguments{
\item{x, y}{Atomic vectors. The length of \code{y} must be greater than the length of \code{x}.}
}
\value{
A tibble of columns \code{.from}, \code{.to} and \code{.n}.
}
\description{
Find missing elements in \code{x} with respect to \code{y}
}
