% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{tile2}
\alias{tile2}
\alias{tile2_lst}
\alias{tile2_dfr}
\alias{ptile}
\alias{ptile_lst}
\alias{ptile_dfr}
\title{Tiling window calculation over multiple inputs simultaneously}
\usage{
tile2(.x, .y, .f, ..., size = 1)

tile2_lst(.x, .y, .f, ..., size = 1)

tile2_dfr(.x, .y, .f, ..., size = 1, .id = NULL)

ptile(.l, .f, ..., size = 1)

ptile_lst(.l, .f, ..., size = 1)

ptile_dfr(.l, .f, ..., size = 1, .id = NULL)
}
\arguments{
\item{.x, .y}{A vector of numerics, or data frame. If a data frame, row-wise rolling
window is performed.}

\item{.f}{A function or one-sided formula using purrr-like syntax. If a
formula, it is converted to a function.}

\item{...}{Additional arguments passed on to \code{.f}.}

\item{size}{An integer for window size.}

\item{.id}{If not \code{NULL} a variable with this name will be created giving
either the name or the index of the data frame, which is passed to
\link[dplyr:bind_rows]{dplyr::bind_rows}.}

\item{.l}{A list of lists. The length of \code{.l} determines the
number of arguments that \code{.f} will be called with. List
names will be used if present.}
}
\description{
Tiling window calculation over multiple inputs simultaneously
}
\examples{
x <- 1:10
y <- 10:1
tile2(x, y, cor, size = 5)
tile2_lst(x, y, ~ cor(.x, .y), size = 5)
}
\seealso{
\link{slide2} for sliding window with overlapping observations;
\link{stretch2} for expanding more observations
}
