% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-tsibble.R, R/yearmonth.R,
%   R/yearquarter.R, R/yearweek.R
\name{tsibble-vctrs}
\alias{tsibble-vctrs}
\alias{vec_ptype2.tbl_ts}
\alias{vec_cast.tbl_ts}
\alias{vec_cast.yearmonth}
\alias{vec_ptype2.yearmonth}
\alias{vec_arith.yearmonth}
\alias{obj_print_data.yearmonth}
\alias{vec_cast.yearquarter}
\alias{vec_ptype2.yearquarter}
\alias{vec_arith.yearquarter}
\alias{obj_print_data.yearquarter}
\alias{vec_cast.yearweek}
\alias{vec_ptype2.yearweek}
\alias{vec_arith.yearweek}
\title{Internal vctrs methods}
\usage{
vec_ptype2.tbl_ts(x, y, ...)

vec_cast.tbl_ts(x, to, ...)

vec_cast.yearmonth(x, to, ...)

vec_ptype2.yearmonth(x, y, ...)

\method{vec_arith}{yearmonth}(op, x, y, ...)

\method{obj_print_data}{yearmonth}(x, ...)

vec_cast.yearquarter(x, to, ...)

vec_ptype2.yearquarter(x, y, ...)

\method{vec_arith}{yearquarter}(op, x, y, ...)

\method{obj_print_data}{yearquarter}(x, ...)

vec_cast.yearweek(x, to, ...)

vec_ptype2.yearweek(x, y, ...)

\method{vec_arith}{yearweek}(op, x, y, ...)
}
\description{
These methods are the extensions that allow tsibble objects to
work with vctrs.
}
\keyword{internal}
