% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsummerise.R
\name{tsummarise}
\alias{tsummarise}
\alias{tsummarize}
\title{Aggregate over calendar periods}
\usage{
tsummarise(.data, ...)

tsummarize(.data, ...)
}
\arguments{
\item{.data}{A tsibble (of \code{tbl_ts} class).}

\item{...}{Name-value pairs of summary functions. To aggregate tsibble over
a certain calendar period, for example yearly aggregates, use \link[lubridate:year]{lubridate::year}
on the index variable.}
}
\value{
A tsibble class.
}
\description{
It computes summary statistics for a tsibble over calendar
periods, usually used in combination of \link{group_by}.
}
\examples{
   # pkgs_ts <- as_tsibble(tidypkgs, index = date, package)
   # pkgs_ts \%>\% 
   #   group_by(package) \%>\% 
   # summarise(avg_count = mean(count), month = ~ as.yearmon())


}
