% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{units_since}
\alias{units_since}
\title{Time units since Unix Epoch}
\usage{
units_since(x)
}
\arguments{
\item{x}{An object of \code{POSIXct}, \code{Date}, \code{yearweek}, \code{yearmonth}, \code{yearquarter}.}
}
\description{
Time units since Unix Epoch
}
\examples{
units_since(x = yearmonth(2012 + (0:11) / 12))
}
