% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyr-verbs.R
\name{unnest_tsibble}
\alias{unnest_tsibble}
\title{Unnest a data frame consisting of tsibbles to a tsibble}
\usage{
unnest_tsibble(data, cols, key = NULL, validate = TRUE)
}
\arguments{
\item{data}{A data frame contains homogenous tsibbles in the list-columns.}

\item{cols}{Names of columns to unnest.}

\item{key}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Variable(s) that uniquely
determine time indices. \code{NULL} for an empty key, unquoted column names (e.g.
\code{x}) for a single variable, and \code{c()} for multiple variables (e.g. \code{c(x, y)}).
This argument also supports \link[tidyselect:language]{tidy-select expressions},
e.g. \code{\link[dplyr:starts_with]{dplyr::starts_with()}}, \code{\link[dplyr:all_of]{dplyr::all_of()}}.}

\item{validate}{\code{TRUE} suggests to verify that each key or each combination
of key variables leads to unique time indices (i.e. a valid tsibble). If you
are sure that it's a valid input, specify \code{FALSE} to skip the checks.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
