% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeartrimester.R
\name{yeartrimester}
\alias{yeartrimester}
\alias{make_yeartrimester}
\alias{is_yeartrimester}
\alias{academic_start}
\title{Represent year-trimester}
\usage{
yeartrimester(x, academic_start = 1)

make_yeartrimester(year = 1970L, trimester = 1L, academic_start = 1)

is_yeartrimester(x)

academic_start(x)
}
\arguments{
\item{x}{Other object.}

\item{year, trimester}{A vector of numerics give years and Trimesters.}
}
\value{
year-trimester (\code{yeartrimester}) objects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Create or coerce using \code{yeartrimester()}.
}
\section{Display}{

Use \code{format()} to display \code{yearweek}, \code{yearmonth}, and \code{yearquarter} objects
in required formats.
Please see \code{\link[=strptime]{strptime()}} details for supported conversion specifications.
}

\examples{
# coerce POSIXct/Dates to yeartrimester
x <- seq(as.Date("2016-01-01"), as.Date("2016-12-31"), by = "1 trimester")
yeartrimester(x)
yeartrimester(x, academic_start = 6)

# parse characters
yeartrimester(c("2018 T1", "2018 tri1", "2018 trimester 1"))

# seq() and arithmetic
tri <- yeartrimester("2017 Q1")
seq(tri, length.out = 10, by = 1) # by 1 trimester
tri + 0:9

# display formats
format(tri, format = "\%y tri\%q")

make_yeartrimester(year = 2021, trimester = 2:3)
make_yeartrimester(year = 2020:2021, trimester = 2:3)

# `academic_start()` helps to extract fiscal year
y <- yeartrimester(as.Date("2020-06-01"), academic_start = 6)
fiscal_year(y)
lubridate::year(y) # calendar years
}
\seealso{
\link{scale_x_yeartrimester} and others for ggplot2 scales

Other index functions: 
\code{\link{yearmonth}()},
\code{\link{yearquarter}()},
\code{\link{yearweek}()}
}
\concept{index functions}
