% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_index}
\alias{aggregate_index}
\alias{aggregate_index.tbl_ts}
\title{Temporal aggregation of the time index}
\usage{
aggregate_index(.data, .granule, ...)

\method{aggregate_index}{tbl_ts}(.data, .granule = NULL, ..., .name = ".granule")
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}

\item{.granule}{A list of mixtime time units to be the linear time chronons of
the aggregated time indices. These time units are evaluated in the context
of the index variable's calendar.}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> Name-value pairs of
summary functions. The name will be the name of the variable in the result.

The value can be:
\itemize{
\item A vector of length 1, e.g. \code{min(x)}, \code{n()}, or \code{sum(is.na(y))}.
\item A data frame with 1 row, to add multiple columns from a single expression.
}}

\item{.name}{The name of the additional key variable identifying the granule
of the aggregated time indices.}

\item{data}{A tsibble.}
}
\description{
Aggregate a tsibble with one or more levels of temporal aggregation. The
granularity of the aggregated time indices is determined by the \code{.granule}
argument, which is the chronons of the aggregated time indices.
}
\examples{
library(tsibble)
pedestrian \%>\% 
  aggregate_index(
    list(hour(12L), day(1L), month(1L), quarter(1L), year(1L)),
    Count = sum(Count)
  )

}
\seealso{
\code{\link[=aggregate_key]{aggregate_key()}}
}
