% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_index}
\alias{aggregate_index}
\title{Temporal aggregation of the time index}
\usage{
aggregate_index(.data, .granule, ..., .name = ".granule")
}
\arguments{
\item{.granule}{A list of mixtime time units to be the linear time chronons of
the aggregated time indices. These time units are evaluated in the context
of the index variable's calendar.}

\item{.name}{The name of the additional key variable identifying the granule
of the aggregated time indices.}

\item{data}{A tsibble.}
}
\description{
Aggregate a tsibble with one or more levels of temporal aggregation. The
granularity of the aggregated time indices is determined by the \code{.granule}
argument, which is the chronons of the aggregated time indices.
}
\examples{
library(tsibble)
pedestrian \%>\% 
  aggregate_index(
    list(hour(12L), day(1L), month(1L), quarter(1L), year(1L)),
    Count = sum(Count)
  )

}
\seealso{
\code{\link[=aggregate_key]{aggregate_key()}}
}
