% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tsibble.R
\name{as_tibble.tbl_ts}
\alias{as_tibble.tbl_ts}
\alias{as.data.frame.tbl_ts}
\title{Coerce to a tibble or data frame}
\usage{
\method{as_tibble}{tbl_ts}(x, ungroup = FALSE, ...)

\method{as.data.frame}{tbl_ts}(x, ...)
}
\arguments{
\item{x}{A \code{tbl_ts}.}

\item{ungroup}{\code{TRUE} to remove the grouping variables.}

\item{...}{Ignored.}
}
\description{
Coerce to a tibble or data frame
}
\examples{
as_tibble(pedestrian)

# a grouped tbl_ts -----
grped_ped <- pedestrian \%>\% group_by(Sensor)
as_tibble(grped_ped)
}
