% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsibble.R
\name{as_tsibble}
\alias{as_tsibble}
\alias{as_tsibble.default}
\alias{as_tsibble.ts}
\alias{as_tsibble.mts}
\alias{as_tsibble.hts}
\alias{as_tsibble.gts}
\title{Coerce to a tsibble object}
\usage{
as_tsibble(x, ...)

\method{as_tsibble}{default}(x, key = key_vars(), index, ...)

\method{as_tsibble}{ts}(x, tz = "UTC", ...)

\method{as_tsibble}{mts}(x, tz = "UTC", ...)

\method{as_tsibble}{hts}(x, tz = "UTC", ...)

\method{as_tsibble}{gts}(x, tz = "UTC", ...)
}
\arguments{
\item{x}{Other objects to be coerced to tsibble.}

\item{...}{Other arguments to be passed.}

\item{key}{Unquoted variable(s) indicating the key variables for tsibble,
used in combination with \code{key_vars()}.}

\item{index}{An unquoted variable indicating the time index variable}

\item{tz}{Time zone.}
}
\value{
A tsibble object.
}
\description{
Coerce to a tsibble object
}
\examples{
   # coerce data.frame to tsibble
   # as_tsibble(tidypkgs, key = key_vars(package), index = date) 

   # coerce ts to tsibble
   as_tsibble(AirPassengers)
   as_tsibble(sunspot.year)
   as_tsibble(sunspot.month)
   as_tsibble(austres)

}
\seealso{
\link[tibble:as_tibble]{tibble::as_tibble}
}
\author{
Earo Wang
}
