% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tsibble.R, R/ts-tsibble.R
\name{as_tsibble}
\alias{as_tsibble}
\alias{as_tsibble.tbl_df}
\alias{as_tsibble.data.frame}
\alias{as_tsibble.tbl}
\alias{as_tsibble.list}
\alias{as_tsibble.grouped_df}
\alias{as_tsibble.default}
\alias{as.tsibble}
\alias{as_tsibble.ts}
\alias{as_tsibble.mts}
\alias{as_tsibble.hts}
\alias{as_tsibble.gts}
\title{Coerce to a tsibble object}
\usage{
as_tsibble(x, ...)

\method{as_tsibble}{tbl_df}(x, ..., index, validate = TRUE)

\method{as_tsibble}{data.frame}(x, ..., index, validate = TRUE)

\method{as_tsibble}{list}(x, ..., index, validate = TRUE)

\method{as_tsibble}{default}(x, ...)

\method{as_tsibble}{ts}(x, tz = "UTC", ...)

\method{as_tsibble}{mts}(x, tz = "UTC", ...)

\method{as_tsibble}{hts}(x, tz = "UTC", ...)

\method{as_tsibble}{gts}(x, tz = "UTC", ...)
}
\arguments{
\item{x}{Other objects to be coerced to tsibble.}

\item{...}{Key variables.}

\item{index}{A bare (or unquoted) variable indicating time index.}

\item{validate}{Logical.}

\item{tz}{Time zone.}
}
\value{
A tsibble object.
}
\description{
Coerce to a tsibble object
}
\examples{
# coerce data.frame to tsibble
df <- data.frame(
  date = rep(seq(as.Date("2017-01-01"), as.Date("2017-01-10"), by = 1), 3),
  group = rep(c("x", "y", "z"), each = 10),
  value = rnorm(30)
)
as_tsibble(df, group) # "date" is automatically considered as the index var
as_tsibble(df, group, index = date) # specify the index var

# coerce ts to tsibble
as_tsibble(AirPassengers)
as_tsibble(sunspot.year)
as_tsibble(sunspot.month)
as_tsibble(austres)

# coerce mts to tsibble
z <- ts(matrix(rnorm(300), 100, 3), start = c(1961, 1), frequency = 12)
as_tsibble(z)

}
