% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tsibble.R
\name{as_tsibble}
\alias{as_tsibble}
\alias{as_tsibble.ts}
\alias{as_tsibble.mts}
\alias{as_tsibble.hts}
\alias{as_tsibble.gts}
\alias{as_tsibble.tbl_df}
\alias{as_tsibble.data.frame}
\alias{as_tsibble.tbl}
\alias{as_tsibble.list}
\alias{key}
\alias{interval}
\alias{index}
\title{Coerce to a tsibble object}
\usage{
as_tsibble(x, ...)

\method{as_tsibble}{ts}(x, tz = "UTC", ...)

\method{as_tsibble}{mts}(x, tz = "UTC", ...)

\method{as_tsibble}{hts}(x, tz = "UTC", ...)

\method{as_tsibble}{gts}(x, tz = "UTC", ...)

\method{as_tsibble}{tbl_df}(x, index, ...)

\method{as_tsibble}{data.frame}(x, index, ...)

\method{as_tsibble}{tbl}(x, index, ...)

\method{as_tsibble}{list}(x, index, ...)

key(x)

interval(x)

index(x)
}
\arguments{
\item{x}{Other objects to be coerced to tsibble.}

\item{...}{Key variables.}

\item{tz}{Time zone.}
}
\value{
A tsibble object.
}
\description{
Coerce to a tsibble object
}
\examples{
   # coerce data.frame to tsibble
   # as_tsibble(tidypkgs, index = date, package) 

   # coerce ts to tsibble
   as_tsibble(AirPassengers)
   as_tsibble(sunspot.year)
   as_tsibble(sunspot.month)
   as_tsibble(austres)

}
\seealso{
\link[tibble:as_tibble]{tibble::as_tibble}
}
\author{
Earo Wang
}
