% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tsibble.R, R/ts-tsibble.R
\name{as_tsibble}
\alias{as_tsibble}
\alias{as.tsibble}
\alias{as_tsibble.tbl_df}
\alias{as_tsibble.data.frame}
\alias{as_tsibble.list}
\alias{as.tsibble}
\alias{as_tsibble.ts}
\alias{as_tsibble.mts}
\alias{as_tsibble.hts}
\title{Coerce to a tsibble object}
\usage{
as_tsibble(x, ...)

\method{as_tsibble}{tbl_df}(x, ..., index, validate = TRUE, regular = TRUE)

\method{as_tsibble}{data.frame}(x, ..., index, validate = TRUE,
  regular = TRUE)

\method{as_tsibble}{list}(x, ..., index, validate = TRUE, regular = TRUE)

\method{as_tsibble}{ts}(x, tz = "UTC", ...)

\method{as_tsibble}{mts}(x, tz = "UTC", gather = TRUE, ...)

\method{as_tsibble}{hts}(x, tz = "UTC", ...)
}
\arguments{
\item{x}{Other objects to be coerced to a tsibble (\code{tbl_ts}).}

\item{...}{Unquoted key variable(s) that define unique time indices. Only used
for data of class \code{tbl_df}, \code{data.frame} and \code{list} (i.e. long data form), if
a univariate time series (without an explicit key), simply leave it blank.
Ignored for other types of classes. See below for details.}

\item{index}{A bare (or unquoted) variable to specify the time index variable.}

\item{validate}{\code{TRUE} suggests to verify that each key or each combination
of key variables lead to unique time indices (i.e. a valid tsibble). It will
also make sure that the nested variables are arranged from lower level to
higher, if nested variables are passed to key (\code{...}). If you are sure that
it's a valid input, specify \code{FALSE} to skip the checks.}

\item{regular}{Regular time interval (\code{TRUE}) or irregular (\code{FALSE}). \code{TRUE}
finds the minimal time span as the interval.}

\item{tz}{Time zone. May be useful when a \code{ts} object is more frequent than
daily.}

\item{gather}{TRUE gives a "long" data form, otherwise as "wide" as \code{x}.}
}
\value{
A tsibble object.
}
\description{
Coerce to a tsibble object
}
\details{
A valid tsibble does not impose the chronological order. Please use
\link{arrange} to get the order by time.
}
\section{Index}{

The time indices are no longer an attribute (for example, the \code{tsp} attribute
in a \code{ts} object), but preserved as the essential component of the tsibble. A
few index class, such as \code{Date}, \code{POSIXt}, and \code{difftime}, forms the basis of
the tsibble, with new additions \link{yearmth} and \link{yearqtr} representing year-month
and year-quarter respectively.
}

\section{Key}{

Key variable(s) together with the index uniquely identifies each record. And
key(s) also imposes the structure on a tsibble to describe other measured
variables:
\itemize{
\item None: an implicit variable
\item A single variable: an explicit variable. For example, \code{data(pedestrian)}
uses the \code{Sensor} column as the key.
\item Nested variables: a nesting of one variable under another. For example,
\code{data(tourism)} contains two geographical locations: \code{Region} and \code{State}.
\code{Region} is the lower level than \code{State} in Australia; in other words, \code{Region}
is nested into \code{State}, which naturally forms a hierarchy. A vertical bars (\code{|})
is used to describe this nesting relationship, and thus \code{Region} | \code{State}.
In theory, nesting can involve infinite levels, so is \code{tsibble}.
\item Crossed variables: a crossing of one variable with another. For example,
the geographical locations are crossed with the purpose of visiting (\code{Purpose})
in the \code{data(tourism)}. A comma (\code{,}) is used to indicate this crossing
relationship. Nested and crossed variables can be combined, such as \code{data(tourism)}.
}

These key variables are descriptors, not values.
}

\examples{
# coerce tibble to tsibble w/o a key ----
tbl1 <- tibble::tibble(
  date = seq(as.Date("2017-01-01"), as.Date("2017-01-10"), by = 1),
  value = rnorm(10)
)
# "date" is automatically considered as the index var, and "group" is the key
as_tsibble(tbl1)
# specify the index var
as_tsibble(tbl1, index = date)

# coerce tibble to tsibble with one key ----
tbl2 <- tibble::tibble(
  date = rep(seq(as.Date("2017-01-01"), as.Date("2017-01-10"), by = 1), 3),
  group = rep(c("x", "y", "z"), each = 10),
  value = rnorm(30)
)
as_tsibble(tbl2, group)
as_tsibble(tbl2, group, index = date)

# coerce ts to tsibble
as_tsibble(AirPassengers)
as_tsibble(sunspot.year)
as_tsibble(sunspot.month)
as_tsibble(austres)

# coerce mts to tsibble
z <- ts(matrix(rnorm(300), 100, 3), start = c(1961, 1), frequency = 12)
as_tsibble(z)
as_tsibble(z, gather = FALSE)

# coerce hts from the "hts" package to tsibble
if (!requireNamespace("hts", quietly = TRUE)) {
  stop("Please install the hts package to run these following examples.")
}
as_tsibble(hts::htseg1)
as_tsibble(hts::htseg2)
}
