% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tsibble.R, R/ts-tsibble.R
\name{as_tsibble}
\alias{as_tsibble}
\alias{as.tsibble}
\alias{as_tsibble.tbl_df}
\alias{as_tsibble.data.frame}
\alias{as_tsibble.list}
\alias{as.tsibble}
\alias{as_tsibble.ts}
\alias{as_tsibble.mts}
\alias{as_tsibble.hts}
\title{Coerce to a tsibble object}
\usage{
as_tsibble(x, ...)

\method{as_tsibble}{tbl_df}(x, ..., index, validate = TRUE, regular = TRUE)

\method{as_tsibble}{data.frame}(x, ..., index, validate = TRUE,
  regular = TRUE)

\method{as_tsibble}{list}(x, ..., index, validate = TRUE, regular = TRUE)

\method{as_tsibble}{ts}(x, tz = "UTC", ...)

\method{as_tsibble}{mts}(x, tz = "UTC", ...)

\method{as_tsibble}{hts}(x, tz = "UTC", ...)
}
\arguments{
\item{x}{Other objects to be coerced to a tsibble (\code{tbl_ts}).}

\item{...}{Unquoted key variable(s) that define unique time indices. Only used
for data of class \code{tbl_df}, \code{data.frame} and \code{list} (i.e. long data form), if
a univariate time series (without an explicit key), simply leave it blank.
Ignored for other types of classes.}

\item{index}{A bare (or unquoted) variable to specify the time index variable.}

\item{validate}{\code{TRUE} suggests to verify that each key or each combination
of key variables lead to unique time indices (i.e. a valid tsibble). If you
are sure that it's a valid input, specify \code{FALSE} to skip the checks.}

\item{regular}{Regular time interval (\code{TRUE}) or irregular (\code{FALSE}). \code{TRUE}
finds the minimal time span as the interval.}

\item{tz}{Time zone. May be useful when a \code{ts} object is more frequent than
daily.}
}
\value{
A tsibble object.
}
\description{
Coerce to a tsibble object
}
\section{Key}{

\strong{Key} informs a tsibble of the structure:
\itemize{
\item Univariate time series: none.
\item Multivariate time series: a single variable. For example, \code{data(pedestrian)}
uses \code{Sensor} as the key that uniquely identify a sequence of time records.
\item Grouped time series: multiple variables separated by commas (\code{,}). For example,
sales data over time can be split by product types (\code{product}) and geographical
locations (\code{geo}), and thus \code{product}, \code{geo} suggests the structure.
\item Hierarchical time series: nested keys using vertical bars (\code{|}). For example,
a natural hierarchy is defined on geographical regions: cities (\code{city}) are
nested in states (\code{state}). The expression of \code{city} | \code{state} is expected.
The general principle is that lower levels are nested in higher levels
(\code{lower} | \code{middle} | \code{higher}).
}

A combination of nested and grouped variables are also supported. Each key or
each combination of keys determine a sequence of unique time indices.
}

\section{Print options}{

The tsibble package fully utilises the \code{print} method from the tibble. Please
refer to \link[tibble:tibble-package]{tibble::tibble-package} to change display options.
}

\examples{
# coerce data.frame to tsibble
df <- data.frame(
  date = rep(seq(as.Date("2017-01-01"), as.Date("2017-01-10"), by = 1), 3),
  group = rep(c("x", "y", "z"), each = 10),
  value = rnorm(30),
  stringsAsFactors = FALSE
)
# "date" is automatically considered as the index var, and "group" is the key
as_tsibble(df, group) 
# specify the index var
as_tsibble(df, group, index = date)

# coerce ts to tsibble
as_tsibble(AirPassengers)
as_tsibble(sunspot.year)
as_tsibble(sunspot.month)
as_tsibble(austres)

# coerce mts to tsibble
z <- ts(matrix(rnorm(300), 100, 3), start = c(1961, 1), frequency = 12)
as_tsibble(z)

# coerce hts from the "hts" package to tsibble
if (!requireNamespace("hts", quietly = TRUE)) {
  stop("Please install the hts package to run these following examples.")
}
as_tsibble(hts::htseg1)
as_tsibble(hts::htseg2)
}
