% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tsibble.R, R/ts-tsibble.R
\name{as_tsibble}
\alias{as_tsibble}
\alias{as_tsibble.tbl_df}
\alias{as_tsibble.data.frame}
\alias{as_tsibble.tbl}
\alias{as_tsibble.list}
\alias{as_tsibble.grouped_df}
\alias{as_tsibble.default}
\alias{key}
\alias{interval}
\alias{index}
\alias{as_tsibble.ts}
\alias{as_tsibble.mts}
\alias{as_tsibble.hts}
\alias{as_tsibble.gts}
\title{Coerce to a tsibble object}
\usage{
as_tsibble(x, ...)

\method{as_tsibble}{tbl_df}(x, index, ..., validate = TRUE)

\method{as_tsibble}{data.frame}(x, index, ..., validate = TRUE)

\method{as_tsibble}{tbl}(x, index, ..., validate = TRUE)

\method{as_tsibble}{list}(x, index, ..., validate = TRUE)

\method{as_tsibble}{grouped_df}(x, index, ..., validate = TRUE)

\method{as_tsibble}{default}(x, index, ...)

key(x)

interval(x)

index(x)

\method{as_tsibble}{ts}(x, tz = "UTC", ...)

\method{as_tsibble}{mts}(x, tz = "UTC", ...)

\method{as_tsibble}{hts}(x, tz = "UTC", ...)

\method{as_tsibble}{gts}(x, tz = "UTC", ...)
}
\arguments{
\item{x}{Other objects to be coerced to tsibble.}

\item{...}{Key variables.}

\item{index}{A bare (or unquoted) variable indicating time index}

\item{validate}{Logical.}

\item{tz}{Time zone.}
}
\value{
A tsibble object.
}
\description{
Coerce to a tsibble object
}
\examples{
   # coerce data.frame to tsibble
   # as_tsibble(tidypkgs, index = date, package) 

# coerce ts to tsibble
as_tsibble(AirPassengers)
as_tsibble(sunspot.year)
as_tsibble(sunspot.month)
as_tsibble(austres)

# coerce mts to tsibble
z <- ts(matrix(rnorm(300), 100, 3), start = c(1961, 1), frequency = 12)
as_tsibble(z)

}
