% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsibble2ts.R
\name{as.ts.tbl_ts}
\alias{as.ts.tbl_ts}
\title{Coerce a tsibble to a time series}
\usage{
\method{as.ts}{tbl_ts}(x, value, frequency = NULL, fill = NA_real_, ...)
}
\arguments{
\item{x}{A \code{tbl_ts} object.}

\item{value}{A measured variable of interest to be spread over columns, if
multiple measures.}

\item{frequency}{A smart frequency with the default \code{NULL}. If set, the
preferred frequency is passed to \code{ts()}.}

\item{fill}{A value to replace missing values.}

\item{...}{Ignored for the function.}
}
\value{
A \code{ts} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
# a monthly series
x1 <- as_tsibble(AirPassengers)
as.ts(x1)
}
