% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tsibble.R
\name{build_tsibble_meta}
\alias{build_tsibble_meta}
\title{Low-level & high-performance constructor for a tsibble object}
\usage{
build_tsibble_meta(
  x,
  key_data = NULL,
  index,
  index2,
  ordered = NULL,
  interval = NULL
)
}
\arguments{
\item{x}{A \code{data.frame}, \code{tbl_df}, \code{tbl_ts}, or other tabular objects.}

\item{key_data}{A data frame containing key variables and \code{.rows}. When a data
frame is supplied, the argument \code{key} will be ignored.}

\item{index, index2}{Quoted variable name.}

\item{ordered}{\code{TRUE} suggests to skip the ordering as \code{x} in
the correct order. \code{FALSE} checks the ordering and may give a warning.}

\item{interval}{\code{TRUE} automatically calculates the interval, and \code{FALSE} for
irregular interval. Use the specified interval via \code{\link[=new_interval]{new_interval()}} as is.}
}
\description{
\lifecycle{experimental}

\code{build_tsibble_meta()} does much less checks than \code{build_tsibble()} for
high performance.
}
\keyword{internal}
