% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-na.R
\name{count_gaps}
\alias{count_gaps}
\title{Count the number of implicit missing observations}
\usage{
count_gaps(x, .full = FALSE)
}
\arguments{
\item{x}{A \code{tbl_ts}.}

\item{.full}{\code{FALSE} to insert \code{NA} for each key within its own period. \code{TRUE}
to fill \code{NA} over the entire time span of the data (a.k.a. fully balanced panel).}
}
\description{
Count the number of implicit missing observations for each key
}
\examples{
count_gaps(pedestrian)
count_gaps(pedestrian, .full = TRUE)
count_gaps(tourism)
}
\seealso{
fill_na
}
