% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaps.R
\name{count_gaps}
\alias{count_gaps}
\title{Count implicit gaps}
\usage{
count_gaps(.data, .full = FALSE, .common = FALSE, ...)
}
\arguments{
\item{.data}{A \code{tbl_ts}.}

\item{.full}{\code{FALSE} to find gaps for each series within its own period.
\code{TRUE} to find gaps over the entire time span of the data.}

\item{.common}{If \code{TRUE}, find common time gaps shared by each series.}

\item{...}{Other arguments passed on to individual methods.}
}
\value{
A tibble contains:
\itemize{
\item the "key" of the \code{tbl_ts}
\item ".from": the starting time point of the gap
\item ".to": the ending time point of the gap
\item ".n": the number of implicit missing observations during the time period
}
}
\description{
Count implicit gaps
}
\seealso{
Other implicit gaps handling: \code{\link{fill_gaps}},
  \code{\link{has_gaps}}, \code{\link{scan_gaps}}
}
\concept{implicit gaps handling}
