% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{as.tsibble}
\alias{as.tsibble}
\alias{fill_na}
\title{Deprecated functions}
\usage{
as.tsibble(x, ...)

fill_na(.data, ..., .full = FALSE)
}
\arguments{
\item{x}{Other objects.}
}
\description{
Deprecated functions
}
\keyword{internal}
