% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tsibble.R
\name{is.tsibble}
\alias{is.tsibble}
\title{Deprecated functions}
\usage{
is.tsibble(x)
}
\arguments{
\item{x}{Other objects.}
}
\description{
\itemize{
\item \code{is.tsibble()} \lifecycle{defunct}
}
}
\keyword{internal}
