% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{split_by}
\alias{split_by}
\alias{case_na}
\alias{as.tsibble}
\alias{fill_na}
\title{Deprecated functions}
\usage{
split_by(x, ...)

case_na(formula)

as.tsibble(x, ...)

fill_na(.data, ..., .full = FALSE)
}
\arguments{
\item{x}{A data frame.}

\item{...}{A list of unquoted variables, separated by commas, to split a dataset.}

\item{formula}{A two-sided formula. The LHS expects a vector containing \code{NA},
and the RHS gives the replacement value.}
}
\description{
Deprecated functions

A thin wrapper of \code{dplyr::case_when()} if there are \code{NA}s
}
\examples{
pedestrian \%>\% 
  split_by(Sensor)
x <- rnorm(10)
x[c(3, 7)] <- NA_real_
case_na(x ~ 10)
case_na(x ~ mean(x, na.rm = TRUE))
}
\seealso{
\link[dplyr:case_when]{dplyr::case_when}
}
\keyword{internal}
