% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{is.tsibble}
\alias{is.tsibble}
\alias{time_unit}
\alias{units_since}
\title{Deprecated functions}
\usage{
is.tsibble(x)

time_unit(x)

units_since(x)
}
\arguments{
\item{x}{Other objects.}

\item{.data}{A tsibble.}

\item{...}{Variables passed to tsibble()/as_tsibble().}
}
\description{
\itemize{
\item \code{is.tsibble()} \lifecycle{defunct}
\item \code{time_unit()} \lifecycle{deprecated}
\item \code{units_since()} \lifecycle{deprecated}
}
}
\keyword{internal}
