% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{is.tsibble}
\alias{is.tsibble}
\alias{fill_na}
\alias{id}
\title{Identifiers used for creating key}
\usage{
is.tsibble(x)

fill_na(.data, ..., .full = FALSE)

id(...)
}
\arguments{
\item{...}{Variables passed to tsibble()/as_tsibble().}
}
\description{
Identifiers used for creating key
}
\keyword{internal}
