% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-wise.R
\name{difference}
\alias{difference}
\title{Lagged Differences in C++}
\usage{
difference(x, lag = 1, differences = 1)
}
\arguments{
\item{x}{A numeric vector.}

\item{lag}{An integer indicating which lag to use.}

\item{differences}{An integer indicating the order of the difference.}
}
\value{
A numeric vector of the same size of \code{x} with padded \code{NA}.
}
\description{
Lagged Differences in C++
}
\examples{
# examples from base
difference(1:10, 2)
difference(1:10, 2, 2)
x <- cumsum(cumsum(1:10))
difference(x, lag = 2)
difference(x, differences = 2)
}
