% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stretch.R
\name{future_stretch}
\alias{future_stretch}
\title{Stretching window in parrallel}
\description{
Multiprocessing equivalents of \code{\link[=slide]{slide()}}, \code{\link[=tile]{tile()}}, \code{\link[=stretch]{stretch()}} prefixed by \code{future_}.
\itemize{
\item Variants for corresponding types: \code{future_*_lgl()}, \code{future_*_int()},
\code{future_*_dbl()}, \code{future_*_chr()}, \code{future_*_dfr()}, \code{future_*_dfc()}.
\item Extra arguments \code{.progress} and \code{.options} for enabling progress bar and the
future specific options to use with the workers.
}
}
\alias{future_stretch}
\alias{future_stretch2}
\alias{future_pstretch}
\alias{future_stretch_lgl}
\alias{future_stretch_chr}
\alias{future_stretch_int}
\alias{future_stretch_dbl}
\alias{future_stretch_dfr}
\alias{future_stretch_dfc}
\alias{future_stretch2_lgl}
\alias{future_stretch2_chr}
\alias{future_stretch2_int}
\alias{future_stretch2_dbl}
\alias{future_stretch2_dfr}
\alias{future_stretch2_dfc}
\alias{future_pstretch_lgl}
\alias{future_pstretch_chr}
\alias{future_pstretch_int}
\alias{future_pstretch_dbl}
\alias{future_pstretch_dfr}
\alias{future_pstretch_dfc}
