% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile.R
\name{future_tile()}
\alias{future_tile()}
\title{Tiling window in parallel}
\description{
\lifecycle{questioning}

\strong{The rolling window family will be deprecated in the future. Please consider
using the \href{https://davisvaughan.github.io/slide}{slide} package.}
Multiprocessing equivalents of \code{\link[=slide]{slide()}}, \code{\link[=tile]{tile()}}, \code{\link[=stretch]{stretch()}} prefixed by \code{future_}.
\itemize{
\item Variants for corresponding types: \verb{future_*_lgl()}, \verb{future_*_int()},
\verb{future_*_dbl()}, \verb{future_*_chr()}, \verb{future_*_dfr()}, \verb{future_*_dfc()}.
\item Extra arguments \code{.progress} and \code{.options} for enabling progress bar and the
future specific options to use with the workers.
}
}
\alias{future_tile}
\alias{future_tile2}
\alias{future_ptile}
\alias{future_tile_lgl}
\alias{future_tile_chr}
\alias{future_tile_int}
\alias{future_tile_dbl}
\alias{future_tile_dfr}
\alias{future_tile_dfc}
\alias{future_tile2_lgl}
\alias{future_tile2_chr}
\alias{future_tile2_int}
\alias{future_tile2_dbl}
\alias{future_tile2_dfr}
\alias{future_tile2_dfc}
\alias{future_ptile_lgl}
\alias{future_ptile_chr}
\alias{future_ptile_int}
\alias{future_ptile_dbl}
\alias{future_ptile_dfr}
\alias{future_ptile_dfc}
