% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-na.R
\name{count_gaps}
\alias{count_gaps}
\alias{gaps}
\title{Count implicit gaps}
\usage{
count_gaps(.data, .full = FALSE)

gaps(x, y)
}
\arguments{
\item{.data}{A \code{tbl_ts}.}

\item{.full}{\code{FALSE} to insert \code{NA} for each key within its own period. \code{TRUE}
to fill \code{NA} over the entire time span of the data (a.k.a. fully balanced panel).}

\item{x, y}{A vector of numbers, dates, or date-times. The length of \code{y} must
be greater than the length of \code{x}.}
}
\value{
A tibble contains:
\itemize{
\item the "key" of the \code{tbl_ts}
\item "from": the starting time point of the gap
\item "end": the ending time point of the gap
\item "n": the implicit missing observations during the time period
}
}
\description{
\code{count_gaps()} counts gaps for a tsibble; \code{gaps()} find where the gaps in \code{x}
with respect to \code{y}.
}
\examples{
count_gaps(pedestrian)
count_gaps(pedestrian, .full = TRUE)
gaps(x = c(1:3, 5:6, 9:10), y = 1:10)
}
\seealso{
fill_na
}
