% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{group_by_key}
\alias{group_by_key}
\title{Group by key variables}
\usage{
group_by_key(.tbl, .funs = list(), ...)
}
\arguments{
\item{.tbl}{A \code{tbl_ts} object.}

\item{.funs}{List of function calls generated by \code{\link[=funs]{funs()}}, or a
character vector of function names, or simply a function.

Bare formulas are passed to \code{\link[rlang:as_function]{rlang::as_function()}} to create
purrr-style lambda functions. Note that these lambda prevent
hybrid evaluation from happening and it is thus more efficient to
supply functions like \code{mean()} directly rather than in a
lambda-formula.}

\item{...}{Additional arguments for the function calls in
\code{.funs}. These are evaluated only once, with \link[rlang:tidy-dots]{tidy
dots} support.}
}
\description{
Group by key variables
}
\examples{
tourism \%>\%
  group_by_key()
}
