% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{group_by_key}
\alias{group_by_key}
\title{Group by key variables}
\usage{
group_by_key(.data, ..., .drop = key_drop_default(.data))
}
\arguments{
\item{.data}{A \code{tbl_ts} object.}

\item{...}{Ignored.}

\item{.drop}{When \code{.drop = TRUE}, empty groups are dropped. See \code{\link[dplyr:group_by_drop_default]{group_by_drop_default()}} for
what the default value is for this argument.}
}
\description{
\lifecycle{stable}
}
\examples{
tourism \%>\%
  group_by_key()
}
