% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tsibble.R
\name{id}
\alias{id}
\title{Identifier to construct structural variables}
\usage{
id(...)
}
\arguments{
\item{...}{Variables passed to tsibble()/as_tsibble().}
}
\description{
Impose a structure to a tsibble
}
\seealso{
\link{tsibble}, \link{as_tsibble}
}
