% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index-by.R
\name{index_by}
\alias{index_by}
\title{Prepare to aggregate over time and update index}
\usage{
index_by(.data, ...)
}
\arguments{
\item{.data}{A \code{tbl_ts}.}

\item{...}{A name-value pair of expression or a bare variable.
The index functions that can be used, but not limited:
\itemize{
\item \link[lubridate:year]{lubridate::year}: yearly aggregation
\item \link{yearquarter}: quarterly aggregation
\item \link{yearmonth}: monthly aggregation
\item \link{yearweek}: weekly aggregation
\item \link{as.Date} or \link[lubridate:as_date]{lubridate::as_date}: daily aggregation
\item \link[lubridate:ceiling_date]{lubridate::ceiling_date}, \link[lubridate:floor_date]{lubridate::floor_date}, or \link[lubridate:round_date]{lubridate::round_date}:
sub-daily aggregation
\item other index functions from other packages
}}
}
\description{
In spirit of \code{group_by()}, \code{index_by()} prepares for a new index with a less
granular period of time. \code{summarise()} will be operated on each group of the
index, in conjunction with \code{index_by()}.
}
\details{
\itemize{
\item A \code{index_by()}-ed tsibble is indicated by \code{*} in the printing output.
\item \code{index_by()} only works with \code{summarise()}, not other verbs.
\item Time index will not be collapsed by \code{summarise.tbl_ts}.
\item The scoped variants of \code{summarise()} only operate on the non-key and
non-index variables.
}
}
\examples{
# Monthly counts across sensors ----
monthly_ped <- pedestrian \%>\% 
  group_by(Sensor) \%>\% 
  index_by(Year_Month = yearmonth(Date_Time)) \%>\%
  summarise(
    Max_Count = max(Count),
    Min_Count = min(Count)
  )
monthly_ped
index(monthly_ped)

# Using existing variable ----
pedestrian \%>\% 
  group_by(Sensor) \%>\% 
  index_by(Date) \%>\%
  summarise(
    Max_Count = max(Count),
    Min_Count = min(Count)
  )

# Annual trips by Region and State ----
tourism \%>\% 
  index_by(Year = lubridate::year(Quarter)) \%>\% 
  group_by(Region | State) \%>\% 
  summarise(Total = sum(Trips))
}
