% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tsibble.R
\name{index}
\alias{index}
\alias{index_var}
\alias{index2}
\alias{index2_var}
\title{Return index variable from a tsibble}
\usage{
index(x)

index_var(x)

index2(x)

index2_var(x)
}
\arguments{
\item{x}{A tsibble object.}
}
\description{
Return index variable from a tsibble
}
\examples{
index(pedestrian)
index_var(pedestrian)
}
