% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-sum.R
\name{index_sum}
\alias{index_sum}
\title{Extensible index type to tsibble}
\usage{
index_sum(x)
}
\arguments{
\item{x}{An object of index type that the tsibble supports.}
}
\description{
S3 method to add an index type support for a tsibble.
}
\details{
This method is primarily used for adding an index type support in
\link{as_tsibble}.
}
\examples{
index_sum(seq(as.Date("2017-01-01"), as.Date("2017-01-10"), by = 1))
}
\seealso{
\link{pull_interval} for obtaining interval for regularly spaced time.
}
