% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index-valid.R
\name{index_valid}
\alias{index_valid}
\title{Add custom index support for a tsibble}
\usage{
index_valid(x)
}
\arguments{
\item{x}{An object of index type that the tsibble supports.}
}
\value{
\code{TRUE}/\code{FALSE} or \code{NA} (unsure)
}
\description{
\lifecycle{stable}

S3 method to add an index type support for a tsibble.
}
\details{
This method is primarily used for adding an index type support in
\link{as_tsibble}.
}
\examples{
index_valid(seq(as.Date("2017-01-01"), as.Date("2017-01-10"), by = 1))
}
\seealso{
\link{interval_pull} for obtaining interval for regularly spaced time.
}
