% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsibble-meta.R
\name{key_data}
\alias{key_data}
\alias{key_rows}
\alias{n_keys}
\title{Keyed data}
\usage{
key_data(.data)

key_rows(.data)
}
\arguments{
\item{.data}{A tsibble}
}
\description{
Keyed data
}
\examples{
key_data(pedestrian)
}
\keyword{internal}
