% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key-group.R
\name{key}
\alias{key}
\alias{key_vars}
\title{Return key variables}
\usage{
key(x)

key_vars(x)
}
\arguments{
\item{x}{A data frame.}
}
\description{
\code{key()} returns a list of symbols; \code{key_vars()} gives a character vector.
}
\examples{
key(pedestrian)
key_vars(pedestrian)

key(tourism)
key_vars(tourism)
}
