% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tsibble.R
\name{key}
\alias{key}
\alias{unkey}
\alias{unkey.tbl_ts}
\alias{key_vars}
\alias{measured_vars}
\alias{measured_vars.tbl_ts}
\title{Return key and measured variables}
\usage{
key(x)

unkey(x)

\method{unkey}{tbl_ts}(x)

key_vars(x)

measured_vars(x)

\method{measured_vars}{tbl_ts}(x)
}
\arguments{
\item{x}{A data frame.}
}
\description{
\code{key()} returns a list of symbols; \code{key_vars()} gives a character vector.
}
\examples{
# A single key for pedestrian data ----
data(pedestrian)
key(pedestrian)
key_vars(pedestrian)
measured_vars(pedestrian)

# Nested and crossed keys for tourism data ----
data(tourism)
key(tourism)
key_vars(tourism)
measured_vars(tourism)
}
