% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-key.R
\name{key}
\alias{key}
\alias{key_vars}
\alias{unkey}
\title{Return key variables}
\usage{
key(x)

key_vars(x)

unkey(x)
}
\arguments{
\item{x}{A data frame.}
}
\description{
\code{key()} returns a list of symbols; \code{key_vars()} gives a character vector.
}
\examples{
# A single key for pedestrian data ----
key(pedestrian)
key_vars(pedestrian)

# Nested and crossed keys for tourism data ----
key(tourism)
key_vars(tourism)
# unkey() only works for a tsibble with 1 key size ----
sx <- pedestrian \%>\% 
  filter(Sensor == "Southern Cross Station")
unkey(sx)
}
