% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key-group.R
\name{key_drop_default}
\alias{key_drop_default}
\title{Default value for .drop argument for key}
\usage{
key_drop_default(.tbl)
}
\arguments{
\item{.tbl}{A data frame}
}
\description{
Default value for .drop argument for key
}
\keyword{internal}
