% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsibble-meta.R
\name{measures}
\alias{measures}
\alias{measured_vars}
\title{Return measured variables}
\usage{
measures(x)

measured_vars(x)
}
\arguments{
\item{x}{A \code{tbl_ts}.}
}
\description{
Return measured variables
}
\examples{
measures(pedestrian)
measures(tourism)

measured_vars(pedestrian)
measured_vars(tourism)
}
