% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{new_interval}
\alias{new_interval}
\title{Create a time interval}
\usage{
new_interval(...)
}
\arguments{
\item{...}{A list of time units to be included in the interval and their
amounts. "year", "quarter", "month", "week", "day", "hour", "minute", "second",
"millisecond", "microsecond", "nanosecond", "unit" are supported.}
}
\value{
an "interval" class
}
\description{
\code{new_interval()} creates an interval object with the specified values.
}
\examples{
new_interval(hour = 1, minute = 30)
}
