% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tsibble.R
\name{new_tsibble}
\alias{new_tsibble}
\title{Create a subclass of a tsibble}
\usage{
new_tsibble(x, ..., class = NULL)
}
\arguments{
\item{x}{A \code{tbl_ts}, required.}

\item{...}{Name-value pairs defining new attributes other than a tsibble.}

\item{class}{Subclasses to assign to the new object, default: none.}
}
\description{
\Sexpr[results=rd, stage=render]{tsibble:::lifecycle("maturing")}
}
