% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{yearmth}
\alias{yearmth}
\alias{yearqtr}
\title{Represent year-month or year-quarter objects}
\usage{
yearmth(x)

yearqtr(x)
}
\arguments{
\item{x}{Other object.}
}
\value{
Year-month (\code{yearmth}) or year-quarter (\code{yearqtr})
objects.
}
\description{
Create or coerce using \code{yearmth()}, or \code{yearqtr()}
}
\details{
It's a known issue that these attributes will be dropped when using
\link{group_by} and \link{mutate} together. It is recommended to \link{ungroup} first, and
then use \link{mutate}.
}
\section{Index functions}{

The tsibble \code{yearmth()} and \code{yearqtr()} functiosn preserve the time zone of
the input \code{x}, contrasting to their zoo counterparts.
}

\examples{
# coerce dates to yearmth, yearqtr ----
x <- seq(as.Date("2016-01-01"), as.Date("2016-12-31"), by = "1 month")
yearmth(x)
yearqtr(x)

# coerce numerics to yearmth, yearqtr ----
yearmth(seq(2010, 2017, by = 1/12))
yearqtr(seq(2010, 2017, by = 1/4))

# coerce yearmonths to yearqtr ----
y <- yearmth(x)
yearqtr(y)
}
\seealso{
\link{pull_interval}
}
