% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{yearmth}
\alias{yearmth}
\alias{yearqtr}
\title{Represent year-month or year-quarter objects}
\usage{
yearmth(x)

yearqtr(x)
}
\arguments{
\item{x}{A vector of date-time, date.}
}
\value{
A year-month (\code{yearmth}) or year-quarter (\code{yearqtr}) object.
}
\description{
Represent year-month or year-quarter objects
}
\examples{
x <- seq(as.Date("2016-01-01"), as.Date("2016-12-31"), by = 30)
yearmth(x)
yearqtr(x)
}
