% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{yearmth}
\alias{yearmth}
\alias{yearqtr}
\alias{year}
\title{Represent year-month, year-quarter and year objects}
\usage{
yearmth(x)

yearqtr(x)

year(x)
}
\arguments{
\item{x}{A vector of date-time, date.}
}
\value{
Year-month (\code{yearmth}), year-quarter (\code{yearqtr}) and year (\code{year})
objects.
}
\description{
Represent year-month, year-quarter and year objects
}
\details{
It's a known issue that these attributes will be dropped when using
\link{group_by} and \link{mutate} together. It is recommended to \link{ungroup} first, and
then use \link{mutate}.
}
\examples{
x <- seq(as.Date("2016-01-01"), as.Date("2016-12-31"), by = 30)
yearmth(x)
yearqtr(x)
year(x)
}
