% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{pull_interval}
\alias{pull_interval}
\title{Pull time interval from a vector and return a class of "interval"}
\usage{
pull_interval(x, duplicated = TRUE)
}
\arguments{
\item{x}{A vector of \code{POSIXt}, \code{Date}, \code{yearmonth}, \code{yearquarter}, \code{difftime},
\code{hms}, \code{integer}, \code{numeric}.}

\item{duplicated}{\code{TRUE} removes the duplicated elements of \code{x} and compute
the minimal time span.}
}
\value{
An "interval" class (a list) includes "year", "quarter", "month",
"week", "day", "hour", "minute", "second", "unit", and other self-defined
interval.
}
\description{
S3 method to pull time interval from various classes, assuming of regularly
spaced time. It finds a minimal time span from \code{x}.
}
\details{
As an S3 method, the \code{pull_interval()} makes a tsibble extensible to
support custom time index.
}
\examples{
x <- seq(as.Date("2017-10-01"), as.Date("2017-10-31"), by = 3)
pull_interval(x, duplicated = FALSE)
}
