% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{summarise}
\alias{reexports}
\alias{summarize}
\alias{reexports}
\alias{filter}
\alias{reexports}
\alias{mutate}
\alias{reexports}
\alias{select}
\alias{reexports}
\alias{group_by}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{as.yearmon}
\alias{reexports}
\alias{as.yearqtr}
\alias{reexports}
\alias{year}
\alias{reexports}
\alias{as_date}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{summarise}}, \code{\link[dplyr]{summarize}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{group_by}}}

  \item{lubridate}{\code{\link[lubridate]{year}}, \code{\link[lubridate]{as_date}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{zoo}{\code{\link[zoo]{as.yearmon}}, \code{\link[zoo]{as.yearqtr}}}
}}

