% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R, R/type-sum.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{as_tibble}
\alias{tibble}
\alias{new_tibble}
\alias{!!}
\alias{!!!}
\alias{type_sum}
\alias{tbl_sum}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{\%>\%}}}

  \item{rlang}{\code{\link[rlang]{!!}}, \code{\link[rlang]{!!!}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}, \code{\link[tibble]{tibble}}, \code{\link[tibble]{new_tibble}}, \code{\link[tibble]{type_sum}}, \code{\link[tibble]{tbl_sum}}}
}}

