% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R, R/yearmonth.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{as_tibble}
\alias{tibble}
\alias{!!}
\alias{!!!}
\alias{union}
\alias{intersect}
\alias{setdiff}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}}

  \item{generics}{\code{\link[generics:setops]{intersect}}, \code{\link[generics:setops]{setdiff}}, \code{\link[generics:setops]{union}}}

  \item{rlang}{\code{\link[rlang:bang-bang]{!!}}, \code{\link[rlang:bang-bang]{!!!}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}, \code{\link[tibble]{tibble}}}
}}

