% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{group_size}
\alias{reexports}
\alias{n_groups}
\alias{reexports}
\alias{group_indices}
\alias{reexports}
\alias{group_vars}
\alias{reexports}
\alias{groups}
\alias{reexports}
\alias{ungroup}
\alias{reexports}
\alias{group_by}
\alias{reexports}
\alias{mutate}
\alias{reexports}
\alias{filter}
\alias{reexports}
\alias{slice}
\alias{reexports}
\alias{select}
\alias{reexports}
\alias{rename}
\alias{reexports}
\alias{arrange}
\alias{reexports}
\alias{summarise}
\alias{reexports}
\alias{summarize}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{glimpse}
\alias{reexports}
\alias{tbl_sum}
\alias{reexports}
\alias{as_tibble}
\alias{reexports}
\alias{as.tibble}
\alias{reexports}
\alias{!!}
\alias{reexports}
\alias{!!!}
\alias{reexports}
\alias{type_sum}
\alias{reexports}
\alias{obj_sum}
\alias{reexports}
\alias{is_vector_s3}
\alias{reexports}
\alias{pillar_shaft}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{group_size}}, \code{\link[dplyr]{n_groups}}, \code{\link[dplyr]{group_indices}}, \code{\link[dplyr]{group_vars}}, \code{\link[dplyr]{groups}}, \code{\link[dplyr]{ungroup}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{summarize}}, \code{\link[dplyr]{\%>\%}}}

  \item{pillar}{\code{\link[pillar]{type_sum}}, \code{\link[pillar]{obj_sum}}, \code{\link[pillar]{is_vector_s3}}, \code{\link[pillar]{pillar_shaft}}}

  \item{rlang}{\code{\link[rlang]{!!}}, \code{\link[rlang]{!!!}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}, \code{\link[tibble]{tbl_sum}}, \code{\link[tibble]{as_tibble}}, \code{\link[tibble]{as.tibble}}}
}}

