% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{group_vars}
\alias{reexports}
\alias{groups}
\alias{reexports}
\alias{group_by}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{as.yearmon}
\alias{reexports}
\alias{as.yearqtr}
\alias{reexports}
\alias{year}
\alias{reexports}
\alias{as_date}
\alias{reexports}
\alias{glimpse}
\alias{reexports}
\alias{tbl_sum}
\alias{reexports}
\alias{type_sum}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{group_vars}}, \code{\link[dplyr]{groups}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{\%>\%}}}

  \item{lubridate}{\code{\link[lubridate]{year}}, \code{\link[lubridate]{as_date}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}, \code{\link[tibble]{tbl_sum}}, \code{\link[tibble]{type_sum}}}

  \item{zoo}{\code{\link[zoo]{as.yearmon}}, \code{\link[zoo]{as.yearqtr}}}
}}

