% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R, R/dplyr-join.R, R/reexports.R,
%   R/type-sum.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{arrange}
\alias{filter}
\alias{slice}
\alias{select}
\alias{rename}
\alias{mutate}
\alias{transmute}
\alias{summarise}
\alias{summarize}
\alias{group_by}
\alias{ungroup}
\alias{distinct}
\alias{left_join}
\alias{right_join}
\alias{inner_join}
\alias{full_join}
\alias{semi_join}
\alias{anti_join}
\alias{\%>\%}
\alias{vars}
\alias{lead}
\alias{lag}
\alias{groups}
\alias{group_vars}
\alias{group_data}
\alias{group_keys}
\alias{group_rows}
\alias{group_by_drop_default}
\alias{as_tibble}
\alias{tibble}
\alias{new_tibble}
\alias{!!}
\alias{!!!}
\alias{type_sum}
\alias{tbl_sum}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{transmute}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{summarize}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{ungroup}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{left_join}}, \code{\link[dplyr]{right_join}}, \code{\link[dplyr]{inner_join}}, \code{\link[dplyr]{full_join}}, \code{\link[dplyr]{semi_join}}, \code{\link[dplyr]{anti_join}}, \code{\link[dplyr]{\%>\%}}, \code{\link[dplyr]{vars}}, \code{\link[dplyr]{lead}}, \code{\link[dplyr]{lag}}, \code{\link[dplyr]{groups}}, \code{\link[dplyr]{group_vars}}, \code{\link[dplyr]{group_data}}, \code{\link[dplyr]{group_keys}}, \code{\link[dplyr]{group_rows}}, \code{\link[dplyr]{group_by_drop_default}}}

  \item{rlang}{\code{\link[rlang]{!!}}, \code{\link[rlang]{!!!}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}, \code{\link[tibble]{tibble}}, \code{\link[tibble]{new_tibble}}, \code{\link[tibble]{type_sum}}, \code{\link[tibble]{tbl_sum}}}
}}

