% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{group_size}
\alias{n_groups}
\alias{group_indices}
\alias{group_vars}
\alias{groups}
\alias{group_data}
\alias{group_rows}
\alias{ungroup}
\alias{group_by}
\alias{mutate}
\alias{filter}
\alias{slice}
\alias{select}
\alias{rename}
\alias{arrange}
\alias{summarise}
\alias{summarize}
\alias{transmute}
\alias{distinct}
\alias{full_join}
\alias{inner_join}
\alias{left_join}
\alias{right_join}
\alias{semi_join}
\alias{anti_join}
\alias{\%>\%}
\alias{gather}
\alias{spread}
\alias{nest}
\alias{unnest}
\alias{fill}
\alias{glimpse}
\alias{tbl_sum}
\alias{as_tibble}
\alias{tibble}
\alias{new_tibble}
\alias{!!}
\alias{!!!}
\alias{type_sum}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{group_size}}, \code{\link[dplyr]{n_groups}}, \code{\link[dplyr]{group_indices}}, \code{\link[dplyr]{group_vars}}, \code{\link[dplyr]{groups}}, \code{\link[dplyr]{group_data}}, \code{\link[dplyr]{group_rows}}, \code{\link[dplyr]{ungroup}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{summarize}}, \code{\link[dplyr]{transmute}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{full_join}}, \code{\link[dplyr]{inner_join}}, \code{\link[dplyr]{left_join}}, \code{\link[dplyr]{right_join}}, \code{\link[dplyr]{semi_join}}, \code{\link[dplyr]{anti_join}}, \code{\link[dplyr]{\%>\%}}}

  \item{rlang}{\code{\link[rlang]{!!}}, \code{\link[rlang]{!!!}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}, \code{\link[tibble]{tbl_sum}}, \code{\link[tibble]{as_tibble}}, \code{\link[tibble]{tibble}}, \code{\link[tibble]{new_tibble}}, \code{\link[tibble]{type_sum}}}

  \item{tidyr}{\code{\link[tidyr]{gather}}, \code{\link[tidyr]{spread}}, \code{\link[tidyr]{nest}}, \code{\link[tidyr]{unnest}}, \code{\link[tidyr]{fill}}}
}}

