% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{group_vars}
\alias{reexports}
\alias{groups}
\alias{reexports}
\alias{ungroup}
\alias{reexports}
\alias{group_by}
\alias{reexports}
\alias{mutate}
\alias{reexports}
\alias{filter}
\alias{reexports}
\alias{slice}
\alias{reexports}
\alias{select}
\alias{reexports}
\alias{rename}
\alias{reexports}
\alias{arrange}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{glimpse}
\alias{reexports}
\alias{tbl_sum}
\alias{reexports}
\alias{type_sum}
\alias{reexports}
\alias{as_tibble}
\alias{reexports}
\alias{as.tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{group_vars}}, \code{\link[dplyr]{groups}}, \code{\link[dplyr]{ungroup}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{\%>\%}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}, \code{\link[tibble]{tbl_sum}}, \code{\link[tibble]{type_sum}}, \code{\link[tibble]{as_tibble}}, \code{\link[tibble]{as.tibble}}}
}}

