% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tsibble.R
\name{is_regular}
\alias{is_regular}
\alias{is.regular}
\alias{is.regular}
\title{If a tsibble is spaced at regular time or not}
\usage{
is_regular(x)

is.regular(x)
}
\arguments{
\item{x}{A tsibble object.}
}
\description{
If a tsibble is spaced at regular time or not
}
\examples{
data(pedestrian)
is_regular(pedestrian)
}
