% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaps.R
\name{scan_gaps}
\alias{scan_gaps}
\title{Scan a tsibble for implicit missing observations}
\usage{
scan_gaps(.data, .full = FALSE)
}
\arguments{
\item{.data}{A tsibble.}

\item{.full}{\code{FALSE} inserts \code{NA} for each keyed unit within its own period. \code{TRUE}
fills \code{NA} over the entire time span of the data (a.k.a. fully balanced panel).
Other options are \code{start()} and \code{end()}, suggesting to pad \code{NA} for the same
\verb{min(<index>)} or \verb{max(<index>)} alignment time across the units.}
}
\description{
Scan a tsibble for implicit missing observations
}
\examples{
scan_gaps(pedestrian)
}
\seealso{
Other implicit gaps handling: 
\code{\link{count_gaps}()},
\code{\link{fill_gaps}()},
\code{\link{has_gaps}()}
}
\concept{implicit gaps handling}
