% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaps.R
\name{scan_gaps}
\alias{scan_gaps}
\title{Scan a tsibble for implicit missing observations}
\usage{
scan_gaps(.data, .full = FALSE, ...)
}
\arguments{
\item{.data}{A \code{tbl_ts}.}

\item{.full}{\code{FALSE} to find gaps for each series within its own period.
\code{TRUE} to find gaps over the entire time span of the data.}

\item{...}{Other arguments passed on to individual methods.}
}
\description{
Scan a tsibble for implicit missing observations
}
\examples{
scan_gaps(pedestrian)
}
\seealso{
Other implicit gaps handling: \code{\link{count_gaps}},
  \code{\link{fill_gaps}}, \code{\link{has_gaps}}
}
\concept{implicit gaps handling}
