% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{slide}
\alias{slide}
\alias{slide.numeric}
\alias{slide.data.frame}
\alias{slider}
\title{Sliding window function}
\usage{
slide(x, .f, ..., size = 1, fill)

\method{slide}{numeric}(x, .f, ..., size = 1, fill = NA_real_)

\method{slide}{data.frame}(x, .f, ..., size = 1, fill = data.frame(),
  deframe = TRUE)

slider(x, size = 1)
}
\arguments{
\item{x}{A vector of numerics, or a data frame.}

\item{.f}{A function, formula, or atomic vector.}

\item{...}{Additional arguments passed on to \code{.f}.}

\item{size}{Window size.}

\item{fill}{A single value to fill \code{NA}.}

\item{deframe}{TRUE a list is returned. FALSE returns a \code{tbl_ts}/\code{tbl_df}/\code{data.frame}.}
}
\description{
Sliding window function
}
\examples{
x <- 1:10
slide(x, sum, size = 3)
slide(x, sum, size = 3, fill = 0)
slide(x, ~ mean(.), size = 2)
}
