% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{slide}
\alias{slide}
\alias{slide.default}
\alias{slide.list}
\title{Sliding window function}
\usage{
slide(x, .f, ..., size = 1, fill)

\method{slide}{default}(x, .f, ..., size = 1, fill = NA_real_)

\method{slide}{list}(x, .f, ..., size = 1, fill = list())
}
\arguments{
\item{x}{A vector of numerics}

\item{.f}{A function, formula, or atomic vector.}

\item{...}{Additional arguments passed on to \code{.f}.}

\item{size}{Window size.}

\item{fill}{A single value to fill \code{NA}.}
}
\value{
A vector of numerics of the same length as \code{x}.
}
\description{
Sliding window function
}
\examples{
x <- 1:10
slide(x, sum, size = 3)
slide(x, sum, size = 3, fill = 0)
slide(x, ~ mean(.), size = 2)
}
