% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{slide}
\alias{slide}
\alias{slide_dbl}
\alias{slide}
\alias{slide_dfr}
\alias{slide_dfc}
\title{Sliding window calculation}
\usage{
slide(.x, .f, ..., .size = 1, .fill = NA)

slide_dbl(.x, .f, ..., .size = 1, .fill = NA)

slide_dfr(.x, .f, ..., .size = 1, .fill = NA, .id = NULL)

slide_dfc(.x, .f, ..., .size = 1, .fill = NA)
}
\arguments{
\item{.x}{An atomic vector. Instead \link{lslide} takes list & data.frame.}

\item{.f}{A function, formula, or atomic vector.

If a \strong{function}, it is used as is.

If a \strong{formula}, e.g. \code{~ .x + 2}, it is converted to a function. There
are three ways to refer to the arguments:
\itemize{
\item For a single argument function, use \code{.}
\item For a two argument function, use \code{.x} and \code{.y}
\item For more arguments, use \code{..1}, \code{..2}, \code{..3} etc
}

This syntax allows you to create very compact anonymous functions.

If \strong{character vector}, \strong{numeric vector}, or \strong{list}, it
is converted to an extractor function. Character vectors index by name
and numeric vectors index by position; use a list to index by position
and name at different levels. Within a list, wrap strings in \code{\link[=get-attr]{get-attr()}}
to extract named attributes. If a component is not present, the value of
\code{.default} will be returned.}

\item{...}{Additional arguments passed on to \code{.f}.}

\item{.size}{An integer for window size.}

\item{.fill}{A single value or data frame to replace \code{NA}.}

\item{.id}{If not \code{NULL} a variable with this name will be created
giving either the name or the index of the data frame.}
}
\description{
Rolling window with overlapping observations:
\itemize{
\item \code{slide()} always returns a list.
\item \code{slide_lgl()}, \code{slide_int()}, \code{slide_dbl()}, \code{slide_chr()} return vectors
of the corresponding type.
\item \code{slide_dfr()} \code{slide_dfc()} return data frames using row-binding & column-binding.
}
}
\details{
The \code{slide()} function attempts to tackle more general problems using
the purrr-like syntax. For some specialist functions like \code{mean} and \code{sum},
you may like to check out for
\href{https://CRAN.R-project.org/package=RcppRoll}{RcppRoll} for faster performance.
}
\examples{
# sliding through a vector ----
x <- 1:10
slide_dbl(x, mean, .size = 3)
slide_dbl(x, mean, .size = 3, .fill = 0)
slide_lgl(x, ~ mean(.) > 1, .size = 3)
}
\seealso{
\itemize{
\item \link{slide2}, \link{pslide}, \link{lslide}
\item \link{tile} for tiling window without overlapping observations
\item \link{stretch} for expanding more observations
}
}
\alias{slide_lgl}
\alias{slide_chr}
\alias{slide_int}
