% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{slide}
\alias{slide}
\alias{slide_dfr}
\alias{slide_dfc}
\title{Sliding window calculation}
\usage{
slide(.x, .f, ..., .size = 1, .fill = NA, .partial = FALSE,
  .align = "right", .bind = FALSE)

slide_dfr(.x, .f, ..., .size = 1, .fill = NA, .partial = FALSE,
  .align = "right", .bind = FALSE, .id = NULL)

slide_dfc(.x, .f, ..., .size = 1, .fill = NA, .partial = FALSE,
  .align = "right", .bind = FALSE)
}
\arguments{
\item{.x}{An object to slide over.}

\item{.f}{A function, formula, or atomic vector.

If a \strong{function}, it is used as is.

If a \strong{formula}, e.g. \code{~ .x + 2}, it is converted to a function. There
are three ways to refer to the arguments:
\itemize{
\item For a single argument function, use \code{.}
\item For a two argument function, use \code{.x} and \code{.y}
\item For more arguments, use \code{..1}, \code{..2}, \code{..3} etc
}

This syntax allows you to create very compact anonymous functions.

If \strong{character vector}, \strong{numeric vector}, or \strong{list}, it
is converted to an extractor function. Character vectors index by name
and numeric vectors index by position; use a list to index by position
and name at different levels. Within a list, wrap strings in \code{\link[=get-attr]{get-attr()}}
to extract named attributes. If a component is not present, the value of
\code{.default} will be returned.}

\item{...}{Additional arguments passed on to \code{.f}.}

\item{.size}{An integer for window size. If positive, moving forward from left
to right; if negative, moving backward (from right to left).}

\item{.fill}{A value to fill at the left of the data range (\code{NA} by default).
\code{NULL} means no filling.}

\item{.partial}{if \code{TRUE}, partial sliding.}

\item{.align}{Align index at the "\strong{r}ight", "\strong{c}entre"/"center", or "\strong{l}eft"
of the window. If \code{.size} is even for center alignment, "centre-right" & "centre-left"
is needed.}

\item{.bind}{If \code{.x} is a list, should \code{.x} be combined before applying \code{.f}?}

\item{.id}{If not \code{NULL} a variable with this name will be created
giving either the name or the index of the data frame.}
}
\description{
Rolling window with overlapping observations:
\itemize{
\item \code{slide()} always returns a list.
\item \code{slide_lgl()}, \code{slide_int()}, \code{slide_dbl()}, \code{slide_chr()} use the same
arguments as \code{slide()}, but return vectors of the corresponding type.
\item \code{slide_dfr()} \code{slide_dfc()} return data frames using row-binding & column-binding.
}
}
\details{
The \code{slide()} function attempts to tackle more general problems using
the purrr-like syntax. For some specialist functions like \code{mean} and \code{sum},
you may like to check out for
\href{https://CRAN.R-project.org/package=RcppRoll}{RcppRoll} for faster performance.
}
\examples{
x <- 1:5
lst <- list(x = x, y = 6:10, z = 11:15)
slide_dbl(x, mean, .size = 2)
slide_dbl(x, mean, .size = 2, align = "center")
slide_lgl(x, ~ mean(.) > 2, .size = 2)
slide(lst, ~ ., .size = 2)
}
\seealso{
\itemize{
\item \link{slide2}, \link{pslide}
\item \link{tile} for tiling window without overlapping observations
\item \link{stretch} for expanding more observations
}
}
\alias{slide_lgl}
\alias{slide_chr}
\alias{slide_int}
\alias{slide_dbl}
