% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{slide_tsibble}
\alias{slide_tsibble}
\title{Perform sliding windows on a tsibble by row}
\usage{
slide_tsibble(.x, .size = 1, .id = ".id")
}
\arguments{
\item{.x}{A tsibble.}

\item{.size}{A positive integer for window size.}

\item{.id}{A character naming the new column \code{.id} containing the partition.}
}
\value{
A tsibble
}
\description{
Perform sliding windows on a tsibble by row
}
\examples{
harvest <- tsibble(
  year = rep(2010:2012, 2),
  fruit = rep(c("kiwi", "cherry"), each = 3),
  kilo = sample(1:10, size = 6),
  key = id(fruit), index = year
)
harvest \%>\% 
  slide_tsibble(.size = 2)
}
\seealso{
Other rolling tsibble: \code{\link{stretch_tsibble}},
  \code{\link{tile_tsibble}}
}
\concept{rolling tsibble}
