% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R, R/stretch.R
\name{slider}
\alias{slider}
\alias{pslider}
\alias{stretcher}
\title{Splits the input to a list according to the rolling window size.}
\usage{
slider(.x, .size = 1)

pslider(..., .size = 1)

stretcher(..., .size = 1, .init = 1)
}
\arguments{
\item{.x}{An object to slide over.}

\item{.size, .init}{An integer for moving and initial window size.}

\item{...}{Multiple objects to be splitted in parallel.}

\item{x}{An objects to be splitted.}
}
\description{
Splits the input to a list according to the rolling window size.
}
\examples{
.x <- 1:5
.y <- 6:10
.z <- 11:15
.lst <- list(x = .x, y = .y, z = .z)
.df <- as.data.frame(.lst)

slider(.x, .size = 2)
slider(.lst, .size = 2)
pslider(.lst, .size = 2)
pslider(list(.x, .y), list(.y))
slider(.df, .size = 2)
pslider(.df, .df, .size = 2)
x <- 1:10
stretcher(x, .init = 3)
y <- 10:1
stretcher(x, y, .init = 3)
}
