% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R, R/stretch.R, R/tile.R
\name{slider}
\alias{slider}
\alias{stretcher}
\alias{tiler}
\title{Splits the input to a list according to the rolling window .size.}
\usage{
slider(..., .size = 1)

stretcher(..., .size = 1, .init = 1)

tiler(..., .size = 1)
}
\arguments{
\item{...}{Objects to be splitted.}

\item{.size, .init}{An integer for moving and initial window size.}
}
\description{
Splits the input to a list according to the rolling window .size.
}
\examples{
x <- 1:10
slider(x, .size = 3)
y <- 10:1
slider(x, y, .size = 3)
x <- 1:10
stretcher(x, .init = 3)
y <- 10:1
stretcher(x, y, .init = 3)
x <- 1:10
tiler(x, .size = 3)
y <- 10:1
tiler(x, y, .size = 3)
}
