% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-by.R
\name{split_by}
\alias{split_by}
\title{Split a data frame into a list of subsets by variables}
\usage{
split_by(x, ...)
}
\arguments{
\item{x}{A data frame.}

\item{...}{Unquoted variables used to split a dataset.}
}
\description{
Split a data frame into a list of subsets by variables
}
\examples{
pedestrian \%>\% 
  split_by(Sensor)
}
