% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{stretch}
\alias{stretch}
\alias{stretch.numeric}
\alias{stretch.data.frame}
\alias{stretcher}
\title{Stretching window function}
\usage{
stretch(x, .f, ..., size = 1, init = 1)

\method{stretch}{numeric}(x, .f, ..., size = 1, init = 1)

\method{stretch}{data.frame}(x, .f, ..., size = 1, init = 1,
  deframe = TRUE)

stretcher(x, size = 1, init = 1)
}
\arguments{
\item{x}{A vector of numerics, or data frame.}

\item{.f}{A function, or formula.}

\item{...}{Additional arguments passed on to \code{.f}.}

\item{size}{Window size.}

\item{init}{Initial window size.}

\item{deframe}{TRUE a list is returned. FALSE returns a data frame.}
}
\description{
Stretching window function
}
\examples{
stretch(1:10, mean, init = 3)
}
