% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{tile}
\alias{tile}
\alias{tile.numeric}
\alias{tile.list}
\alias{tile.data.frame}
\title{Tiling window function}
\usage{
tile(x, .f, ..., size = 1)

\method{tile}{numeric}(x, .f, ..., size = 1)

\method{tile}{list}(x, .f, ..., size = 1)

\method{tile}{data.frame}(x, .f, ..., size = 1, deframe = TRUE)
}
\arguments{
\item{x}{A vector of numerics}

\item{.f}{A function, formula, or atomic vector.}

\item{...}{Additional arguments passed on to \code{.f}.}

\item{size}{Window size.}

\item{deframe}{TRUE a list is returned. FALSE returns a \code{tbl_ts}/\code{tbl_df}/\code{data.frame}.}
}
\value{
A vector of numerics of the same length as \code{x}.
}
\description{
Tiling window function
}
\examples{
tile(1:10, mean, size = 3)
}
