% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{tile}
\alias{tile}
\alias{tile.numeric}
\alias{tile.data.frame}
\alias{tiler}
\title{Tiling window function}
\usage{
tile(x, .f, ..., size = 1)

\method{tile}{numeric}(x, .f, ..., size = 1)

\method{tile}{data.frame}(x, .f, ..., size = 1, deframe = TRUE)

tiler(x, size = 1)
}
\arguments{
\item{x}{A vector of numerics, or a data frame.}

\item{.f}{A function, formula, or atomic vector.}

\item{...}{Additional arguments passed on to \code{.f}.}

\item{size}{Window size.}

\item{deframe}{TRUE a list is returned. FALSE returns a \code{tbl_ts}/\code{tbl_df}/\code{data.frame}.}
}
\description{
Tiling window function
}
\examples{
tile(1:10, mean, size = 3)
}
