% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile.R
\name{tile2}
\alias{tile2}
\alias{tile2}
\alias{tile2_dfr}
\alias{tile2_dfc}
\alias{ptile}
\alias{ptile}
\alias{ptile_dfr}
\alias{ptile_dfc}
\title{Tiling window calculation over multiple inputs simultaneously}
\usage{
tile2(.x, .y, .f, ..., .size = 1)

tile2_dfr(.x, .y, .f, ..., .size = 1, .id = NULL)

tile2_dfc(.x, .y, .f, ..., .size = 1)

ptile(.l, .f, ..., .size = 1)

ptile_dfr(.l, .f, ..., .size = 1, .id = NULL)

ptile_dfc(.l, .f, ..., .size = 1)
}
\arguments{
\item{.x}{Objects to slide over simultaneously.}

\item{.y}{Objects to slide over simultaneously.}

\item{.f}{A function, formula, or atomic vector.

If a \strong{function}, it is used as is.

If a \strong{formula}, e.g. \code{~ .x + 2}, it is converted to a function. There
are three ways to refer to the arguments:
\itemize{
\item For a single argument function, use \code{.}
\item For a two argument function, use \code{.x} and \code{.y}
\item For more arguments, use \code{..1}, \code{..2}, \code{..3} etc
}

This syntax allows you to create very compact anonymous functions.

If \strong{character vector}, \strong{numeric vector}, or \strong{list}, it
is converted to an extractor function. Character vectors index by name
and numeric vectors index by position; use a list to index by position
and name at different levels. Within a list, wrap strings in \code{\link[=get-attr]{get-attr()}}
to extract named attributes. If a component is not present, the value of
\code{.default} will be returned.}

\item{...}{Additional arguments passed on to \code{.f}.}

\item{.size}{An integer for window size. If positive, moving forward from left
to right; if negative, moving backward (from right to left).}

\item{.id}{If not \code{NULL} a variable with this name will be created
giving either the name or the index of the data frame.}

\item{.l}{A list of lists. The length of \code{.l} determines the
number of arguments that \code{.f} will be called with. List
names will be used if present.}
}
\description{
Tiling window without overlapping observations:
\itemize{
\item \code{tile2()} and \code{ptile()} always returns a list.
\item \code{tile2_lgl()}, \code{tile2_int()}, \code{tile2_dbl()}, \code{tile2_chr()} use the same
arguments as \code{tile2()}, but return vectors of the corresponding type.
\item \code{tile2_dfr()} \code{tile2_dfc()} return data frames using row-binding & column-binding.
}
}
\examples{
.x <- 1:5
.y <- 6:10
.z <- 11:15
.lst <- list(x = .x, y = .y, z = .z)
.df <- as.data.frame(.lst)
tile2(.x, .y, sum, .size = 2)
tile2(.lst, .lst, sum, .size = 2)
tile2(.df, .df, sum, .size = 2)
ptile(.lst, sum, size = 1)
ptile(list(.lst, .lst), ~ ., .size = 2)
}
\seealso{
\itemize{
\item \link{tile}
\item \link{slide2} for sliding window with overlapping observations
\item \link{stretch2} for expanding more observations
}
}
\alias{tile2_lgl}
\alias{tile2_chr}
\alias{tile2_dbl}
\alias{tile2_int}
\alias{ptile_lgl}
\alias{ptile_chr}
\alias{ptile_dbl}
\alias{ptile_int}
