% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile.R
\name{tile_tsibble}
\alias{tile_tsibble}
\title{Perform tiling windows on a tsibble by row}
\usage{
tile_tsibble(.x, .size = 1)
}
\arguments{
\item{.x}{A tsibble.}

\item{.size}{A positive integer for window size.}
}
\description{
Perform tiling windows on a tsibble by row
}
\section{Rolling tsibble}{

\code{slide_tsibble()}, \code{tile_tsibble()}, and \code{stretch_tsibble()} provide fast
and shorthand for rolling over a tsibble by observations. That said, if the
supplied tsibble has time gaps, these rolling helpers will ignore those gaps
and proceed.

They are useful for preparing the tsibble for time series cross validation.
The output dimension will increase considerably with \code{slide_tsibble()} and
\code{stretch_tsibble()}, which is likely to run out of memory when the data is
large. Alternatively, you could construct cross validation using \code{pslide()}
and \code{pstretch()} to avoid the memory issue.
}

\examples{
harvest <- tsibble(
  year = rep(2010:2012, 2),
  fruit = rep(c("kiwi", "cherry"), each = 3),
  kilo = sample(1:10, size = 6),
  key = fruit, index = year
)
harvest \%>\%
  tile_tsibble(.size = 2)
}
\seealso{
Other rolling tsibble: \code{\link{slide_tsibble}},
  \code{\link{stretch_tsibble}}
}
\concept{rolling tsibble}
