% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{time_unit}
\alias{time_unit}
\title{Extract time unit from a vector}
\usage{
time_unit(x)
}
\arguments{
\item{x}{A vector of \code{POSIXt}, \code{Date}, \code{yearmonth}, \code{yearquarter}, \code{difftime},
\code{hms}, \code{integer}, \code{numeric}.}
}
\description{
Extract time unit from a vector
}
\examples{
x <- yearmonth(seq(2016, 2018, by = 0.5))
time_unit(x)
}
