% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsibble-pkg.R
\docType{package}
\name{tsibble-package}
\title{tsibble: Temporal Data Frames}
\description{
The 'tsibble' provides and works with tidy temporal data.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://pkg.earo.me/tsibble}
  \item Report bugs at \url{https://github.com/earowang/tsibble/issues}
}

}
\author{
\strong{Maintainer}: Earo Wang \email{earo.wang@gmail.com}

}
