% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{tsibble-scales}
\alias{tsibble-scales}
\alias{scale_x_yearquarter}
\alias{scale_x_yearmonth}
\alias{scale_x_yearweek}
\title{tsibble scales for ggplot2}
\usage{
scale_x_yearquarter(...)

scale_x_yearmonth(...)

scale_x_yearweek(...)
}
\arguments{
\item{...}{Further arguments to be passed on to scale_x_date()}
}
\value{
A ggproto object inheriting from \code{Scale}
}
\description{
Defines ggplot2 scales for tsibble time structures.
}
