% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsibble.R
\name{tsibble}
\alias{tsibble}
\alias{key_vars}
\title{Create a tsibble object}
\usage{
tsibble(..., key = key_vars(), index)

key_vars(...)
}
\arguments{
\item{...}{A set of name-value pairs.}

\item{key}{Unquoted variable(s) indicating the key variables for tsibble,
used in combination with \code{key_vars()}.}

\item{index}{An unquoted variable indicating the time index variable.}

\item{...}{Unquoted variable(s).}
}
\value{
A tsibble object.
}
\description{
Create a tsibble object
}
\examples{
   ts_df <- tsibble(
     Date = rep(seq(as.Date("2017-01-01"), by = 1, length = 10), 2),
     Group = rep(c("A", "B"), each = 10),
     Value = rnorm(20),
     key = key_vars(Group), index = Date
   ) 
   print(ts_df)

}
\seealso{
\link[tibble:tibble]{tibble::tibble}
}
\author{
Earo Wang
}
\keyword{internal}
