% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsummarise.R
\name{tsummarise}
\alias{tsummarise}
\alias{tsummarize}
\title{Aggregate over calendar periods}
\usage{
tsummarise(.data, ...)

tsummarize(.data, ...)
}
\arguments{
\item{.data}{A tsibble (of \code{tbl_ts} class).}

\item{...}{Name-value pairs of summary functions. To aggregate the data over
a certain calendar period, for example monthly aggregates, use \link{yearmth}
on the index variable.}
}
\description{
It computes summary statistics for a tsibble over calendar periods, usually
used in combination of \link{group_by}.
}
\examples{
monthly_ped <- pedestrian \%>\% 
  group_by(Sensor) \%>\% 
  tsummarise(
    Year_Month = yearmth(Date_Time), # Year_Month will be the new index
    Max_Count = max(Count),
    Min_Count = min(Count)
  )
monthly_ped
index(monthly_ped)


}
