% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{units_since}
\alias{units_since}
\title{Time units since Unix Epoch}
\usage{
units_since(x)
}
\arguments{
\item{x}{An object of \code{POSIXct}, \code{Date}, \code{yearweek}, \code{yearmonth}, \code{yearquarter}.}
}
\description{
\lifecycle{questioning}
}
\details{
origin:
\itemize{
\item \code{POSIXct}: 1970-01-01 00:00:00
\item \code{Date}: 1970-01-01
\item \code{yearweek}: 1970 W01 (i.e. 1969-12-29)
\item \code{yearmonth}: 1970 Jan
\item \code{yearquarter}: 1970 Qtr1
}
}
\examples{
units_since(x = yearmonth(2012 + (0:11) / 12))
}
