% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-vctrs.R, R/yearmonth.R
\name{vctrs-compat}
\alias{vctrs-compat}
\alias{vec_arith.yearmonth}
\title{vctrs compatibility functions}
\usage{
\method{vec_arith}{yearmonth}(op, x, y, ...)
}
\arguments{
\item{x}{Objects.}

\item{...}{Used to pass along error message information.}

\item{to}{Type to cast to.}
}
\description{
These functions are the extensions that allow mists objects to
work with vctrs.
}
\keyword{internal}
