% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearmonth.R
\name{yearmonth}
\alias{yearmonth}
\alias{is_yearmonth}
\title{Represent year-month}
\usage{
yearmonth(x)

is_yearmonth(x)
}
\arguments{
\item{x}{Other object.}
}
\value{
year-month (\code{yearmonth}) objects.
}
\description{
\lifecycle{stable}

Create or coerce using \code{yearmonth()}.
}
\examples{
# coerce POSIXct/Dates to yearmonth
x <- seq(as.Date("2016-01-01"), as.Date("2016-12-31"), by = "1 month")
yearmonth(x)

# parse characters
yearmonth(c("2018 Jan", "2018-01", "2018 January"))

# creat an empty yearmonth container
yearmonth()

# seq() and arithmetic
mth <- yearmonth("2017-11")
seq(mth, length.out = 10, by = 1) # by 1 month
mth + 0:9

# display formats
format(mth, format = "\%y \%m")
}
\seealso{
\link{interval_pull}, \link{units_since}
}
